Struct vulkano::buffer::cpu_access::CpuAccessibleBuffer

pub struct CpuAccessibleBuffer<T: ?Sized, A = StdMemoryPool> where A: MemoryPool {
    // some fields omitted
}

Buffer whose content is accessible by the CPU.

Methods

impl<T> CpuAccessibleBuffer<T>

fn new<'a, I>(device: &Arc<Device>, usage: &Usage, queue_families: I) -> Result<Arc<CpuAccessibleBuffer<T>>, OomError> where I: IntoIterator<Item=QueueFamily<'a>>

Builds a new buffer. Only allowed for sized data.

impl<T> CpuAccessibleBuffer<[T]>

fn array<'a, I>(device: &Arc<Device>, len: usize, usage: &Usage, queue_families: I) -> Result<Arc<CpuAccessibleBuffer<[T]>>, OomError> where I: IntoIterator<Item=QueueFamily<'a>>

Builds a new buffer. Can be used for arrays.

impl<T: ?Sized> CpuAccessibleBuffer<T>

unsafe fn raw<'a, I>(device: &Arc<Device>, size: usize, usage: &Usage, queue_families: I) -> Result<Arc<CpuAccessibleBuffer<T>>, OomError> where I: IntoIterator<Item=QueueFamily<'a>>

Builds a new buffer without checking the size.

Safety

You must ensure that the size that you pass is correct for T.

impl<T: ?Sized, A> CpuAccessibleBuffer<T, A> where A: MemoryPool

fn device(&self) -> &Arc<Device>

Returns the device used to create this buffer.

fn queue_families(&self) -> Vec<QueueFamily>

Returns the queue families this buffer can be used on.

impl<T: ?Sized, A> CpuAccessibleBuffer<T, A> where T: Content + 'static, A: MemoryPool

fn read(&self, timeout: Duration) -> Result<ReadLock<T>, FenceWaitError>

Locks the buffer in order to write its content.

If the buffer is currently in use by the GPU, this function will block until either the buffer is available or the timeout is reached. A value of 0 for the timeout is valid and means that the function should never block.

After this function successfully locks the buffer, any attempt to submit a command buffer that uses it will block until you unlock it.

fn write(&self, timeout: Duration) -> Result<WriteLock<T>, FenceWaitError>

Locks the buffer in order to write its content.

If the buffer is currently in use by the GPU, this function will block until either the buffer is available or the timeout is reached. A value of 0 for the timeout is valid and means that the function should never block.

After this function successfully locks the buffer, any attempt to submit a command buffer that uses it will block until you unlock it.