Struct vulkano::buffer::cpu_access::ReadLock

pub struct ReadLock<'a, T: ?Sized + 'a> {
    // some fields omitted
}

Object that can be used to read or write the content of a CpuAccessBuffer.

Note that this object holds a rwlock read guard on the chunk. If another thread tries to access this buffer's content or tries to submit a GPU command that uses this buffer, it will block.

Methods

impl<'a, T: ?Sized + 'a> ReadLock<'a, T>

fn map<U: ?Sized + 'a, F>(self, f: F) -> ReadLock<'a, U> where F: FnOnce(&mut T) -> &mut U

Makes a new ReadLock to access a sub-part of the current ReadLock.