Module vulkano::buffer::device_local

Buffer whose content is read-written by the GPU only.

Each access from the CPU or from the GPU locks the whole buffer for either reading or writing. You can read the buffer multiple times simultaneously from multiple queues. Trying to read and write simultaneously, or write and write simultaneously will block with a semaphore.

Structs

DeviceLocalBuffer
Buffer whose content is in device-local memory.