Struct vulkano::buffer::device_local::DeviceLocalBuffer

pub struct DeviceLocalBuffer<T: ?Sized, A = PotentialDedicatedAllocation<StdMemoryPoolAlloc>> { /* fields omitted */ }

Buffer whose content is in device-local memory.

This buffer type is useful in order to store intermediary data. For example you execute a compute shader that writes to this buffer, then read the content of the buffer in a following compute or graphics pipeline.

The DeviceLocalBuffer will be in device-local memory, unless the device doesn't provide any device-local memory.

Methods

impl<T> DeviceLocalBuffer<T>

pub fn new<'a, I>(
device: Arc<Device>,
usage: BufferUsage,
queue_families: I
) -> Result<Arc<DeviceLocalBuffer<T>>, DeviceMemoryAllocError> where
I: IntoIterator<Item = QueueFamily<'a>>,

Builds a new buffer. Only allowed for sized data.

impl<T> DeviceLocalBuffer<[T]>

pub fn array<'a, I>(
device: Arc<Device>,
len: usize,
usage: BufferUsage,
queue_families: I
) -> Result<Arc<DeviceLocalBuffer<[T]>>, DeviceMemoryAllocError> where
I: IntoIterator<Item = QueueFamily<'a>>,

Builds a new buffer. Can be used for arrays.

impl<T: ?Sized> DeviceLocalBuffer<T>

pub unsafe fn raw<'a, I>(
device: Arc<Device>,
size: usize,
usage: BufferUsage,
queue_families: I
) -> Result<Arc<DeviceLocalBuffer<T>>, DeviceMemoryAllocError> where
I: IntoIterator<Item = QueueFamily<'a>>,

Builds a new buffer without checking the size.

Safety

You must ensure that the size that you pass is correct for T.

impl<T: ?Sized, A> DeviceLocalBuffer<T, A>

pub fn queue_families(&self) -> Vec<QueueFamily>

Returns the queue families this buffer can be used on.