Struct vulkano::buffer::immutable::ImmutableBuffer

pub struct ImmutableBuffer<T: ?Sized, A = PotentialDedicatedAllocation<StdMemoryPoolAlloc>> { /* fields omitted */ }

Buffer that is written once then read for as long as it is alive.

Methods

impl<T: ?Sized> ImmutableBuffer<T>

pub fn from_data(
data: T,
usage: BufferUsage,
queue: Arc<Queue>
) -> Result<(Arc<ImmutableBuffer<T>>, CommandBufferExecFuture<NowFuture, AutoCommandBuffer>), DeviceMemoryAllocError> where
T: 'static + Send + Sync + Sized,

Builds an ImmutableBuffer from some data.

This function builds a memory-mapped intermediate buffer, writes the data to it, builds a command buffer that copies from this intermediate buffer to the final buffer, and finally submits the command buffer as a future.

This function returns two objects: the newly-created buffer, and a future representing the initial upload operation. In order to be allowed to use the ImmutableBuffer, you must either submit your operation after this future, or execute this future and wait for it to be finished before submitting your own operation.

pub fn from_buffer<B>(
source: B,
usage: BufferUsage,
queue: Arc<Queue>
) -> Result<(Arc<ImmutableBuffer<T>>, CommandBufferExecFuture<NowFuture, AutoCommandBuffer>), DeviceMemoryAllocError> where
B: BufferAccess + TypedBufferAccess<Content = T> + 'static + Clone + Send + Sync,
T: 'static + Send + Sync,

Builds an ImmutableBuffer that copies its data from another buffer.

This function returns two objects: the newly-created buffer, and a future representing the initial upload operation. In order to be allowed to use the ImmutableBuffer, you must either submit your operation after this future, or execute this future and wait for it to be finished before submitting your own operation.

impl<T> ImmutableBuffer<T>

pub unsafe fn uninitialized(
device: Arc<Device>,
usage: BufferUsage
) -> Result<(Arc<ImmutableBuffer<T>>, ImmutableBufferInitialization<T>), DeviceMemoryAllocError>

Builds a new buffer with uninitialized data. Only allowed for sized data.

Returns two things: the buffer, and a special access that should be used for the initial upload to the buffer.

You will get an error if you try to use the buffer before using the initial upload access. However this function doesn't check whether you actually used this initial upload to fill the buffer like you're supposed to do.

You will also get an error if you try to get exclusive access to the final buffer.

Safety

The ImmutableBufferInitialization should be used to fill the buffer with some initial data, otherwise the content is undefined.

impl<T> ImmutableBuffer<[T]>

pub fn from_iter<D>(
data: D,
usage: BufferUsage,
queue: Arc<Queue>
) -> Result<(Arc<ImmutableBuffer<[T]>>, CommandBufferExecFuture<NowFuture, AutoCommandBuffer>), DeviceMemoryAllocError> where
D: ExactSizeIterator<Item = T>,
T: 'static + Send + Sync + Sized,

pub unsafe fn uninitialized_array(
device: Arc<Device>,
len: usize,
usage: BufferUsage
) -> Result<(Arc<ImmutableBuffer<[T]>>, ImmutableBufferInitialization<[T]>), DeviceMemoryAllocError>

Builds a new buffer with uninitialized data. Can be used for arrays.

Returns two things: the buffer, and a special access that should be used for the initial upload to the buffer.

You will get an error if you try to use the buffer before using the initial upload access. However this function doesn't check whether you actually used this initial upload to fill the buffer like you're supposed to do.

You will also get an error if you try to get exclusive access to the final buffer.

Safety

The ImmutableBufferInitialization should be used to fill the buffer with some initial data, otherwise the content is undefined.

impl<T: ?Sized> ImmutableBuffer<T>

pub unsafe fn raw<'a, I>(
device: Arc<Device>,
size: usize,
usage: BufferUsage,
queue_families: I
) -> Result<(Arc<ImmutableBuffer<T>>, ImmutableBufferInitialization<T>), DeviceMemoryAllocError> where
I: IntoIterator<Item = QueueFamily<'a>>,

Builds a new buffer without checking the size and granting free access for the initial upload.

Returns two things: the buffer, and a special access that should be used for the initial upload to the buffer. You will get an error if you try to use the buffer before using the initial upload access. However this function doesn't check whether you used this initial upload to fill the buffer. You will also get an error if you try to get exclusive access to the final buffer.

Safety

You must ensure that the size that you pass is correct for T.

The ImmutableBufferInitialization should be used to fill the buffer with some initial data.

impl<T: ?Sized, A> ImmutableBuffer<T, A>

pub fn device(&self) -> &Arc<Device>

Returns the device used to create this buffer.

pub fn queue_families(&self) -> Vec<QueueFamily>

Returns the queue families this buffer can be used on.