Struct vulkano::buffer::Buffer

pub struct Buffer<T: ?Sized, M> {
    // some fields omitted
}

Methods

impl<T, M> Buffer<T, M> where M: MemorySourceChunk

fn new<S, Sh>(device: &Arc<Device>, usage: &Usage, memory: S, sharing: Sh) -> Result<Arc<Buffer<T, M>>, OomError> where S: MemorySource<Chunk=M>, Sh: Into<SharingMode>

Creates a new buffer.

impl<T, M> Buffer<[T], M> where M: MemorySourceChunk

fn array<S, Sh>(device: &Arc<Device>, len: usize, usage: &Usage, memory: S, sharing: Sh) -> Result<Arc<Buffer<[T], M>>, OomError> where S: MemorySource<Chunk=M>, Sh: Into<SharingMode>

Creates a new buffer with a number of elements.

impl<T: ?Sized, M> Buffer<T, M> where M: MemorySourceChunk

unsafe fn raw<S, Sh>(device: &Arc<Device>, size: usize, usage: &Usage, memory: S, sharing: Sh) -> Result<Arc<Buffer<T, M>>, OomError> where S: MemorySource<Chunk=M>, Sh: Into<SharingMode>

Creates a new buffer of the given size without checking whether the type is correct.

Safety

Type safety is not checked.

impl<T: ?Sized, M> Buffer<T, M>

fn device(&self) -> &Arc<Device>

Returns the device used to create this buffer.

fn size(&self) -> usize

Returns the size of the buffer in bytes.

impl<T, M> Buffer<[T], M>

fn len(&self) -> usize

Returns the number of elements in the buffer.

impl<'a, T: ?Sized, M> Buffer<T, M> where M: CpuAccessible<'a, T>

fn read(&'a self, timeout_ns: u64) -> M::Read

Gives a read access to the content of the buffer.

If the buffer is in use by the GPU, blocks until it is available.

fn try_read(&'a self) -> Option<M::Read>

Tries to give a read access to the content of the buffer.

If the buffer is in use by the GPU, returns None.

impl<'a, T: ?Sized, M> Buffer<T, M> where M: CpuWriteAccessible<'a, T>

fn write(&'a self, timeout_ns: u64) -> M::Write

Gives a write access to the content of the buffer.

If the buffer is in use by the GPU, blocks until it is available.

fn try_write(&'a self) -> Option<M::Write>

Tries to give a write access to the content of the buffer.

If the buffer is in use by the GPU, returns None.