Struct vulkano::buffer::BufferSlice

pub struct BufferSlice<T: ?Sized, B> { /* fields omitted */ }

A subpart of a buffer.

This object doesn't correspond to any Vulkan object. It exists for API convenience.

Example

Creating a slice:

use vulkano::buffer::BufferSlice;
let _slice = BufferSlice::from(&buffer);

Selecting a slice of a buffer that contains [T]:

use vulkano::buffer::BufferSlice;
let _slice = BufferSlice::from(&buffer).slice(12 .. 14).unwrap();

Methods

impl<T: ?Sized, B> BufferSlice<T, B>

pub fn from_typed_buffer_access(r: B) -> BufferSlice<T, B> where
B: TypedBufferAccess<Content = T>,

pub fn buffer(&self) -> &B

Returns the buffer that this slice belongs to.

pub fn offset(&self) -> usize

Returns the offset of that slice within the buffer.

pub fn size(&self) -> usize

Returns the size of that slice in bytes.

pub unsafe fn slice_custom<F, R: ?Sized>(self, f: F) -> BufferSlice<R, B> where
F: for<'r> FnOnce(&'r T) -> &'r R,

Builds a slice that contains an element from inside the buffer.

This method builds an object that represents a slice of the buffer. No actual operation is performed.

Example

TODO

Safety

The object whose reference is passed to the closure is uninitialized. Therefore you must not access the content of the object.

You must return a reference to an element from the parameter. The closure must not panic.

pub unsafe fn reinterpret<R: ?Sized>(self) -> BufferSlice<R, B>

Changes the T generic parameter of the BufferSlice to the desired type. This can be useful when you have a buffer with various types of data and want to create a typed slice of a region that contains a single type of data.

Example

let blob_slice: BufferSlice<[u8], Arc<ImmutableBuffer<[u8]>>> = return;
let vertex_slice: BufferSlice<[VertexImpl], Arc<ImmutableBuffer<[u8]>>> = unsafe {
    blob_slice.reinterpret::<[VertexImpl]>()
};

Safety

Correct offset and size must be ensured before using this BufferSlice on the device. See BufferSlice::slice for adjusting these properties.

impl<T, B> BufferSlice<[T], B>

pub fn len(&self) -> usize

Returns the number of elements in this slice.

pub fn index(self, index: usize) -> Option<BufferSlice<T, B>>

Reduces the slice to just one element of the array.

Returns None if out of range.

pub fn slice(self, range: Range<usize>) -> Option<BufferSlice<[T], B>>

Reduces the slice to just a range of the array.

Returns None if out of range.