Struct vulkano::buffer::sys::Usage

pub struct Usage {
    pub transfer_source: bool,
    pub transfer_dest: bool,
    pub uniform_texel_buffer: bool,
    pub storage_texel_buffer: bool,
    pub uniform_buffer: bool,
    pub storage_buffer: bool,
    pub index_buffer: bool,
    pub vertex_buffer: bool,
    pub indirect_buffer: bool,
}

Describes how a buffer is going to be used. This is not an optimization.

If you try to use a buffer in a way that you didn't declare, a panic will happen.

Some methods are provided to build Usage structs for some common situations. However there is no restriction in the combination of usages that can be enabled.

Fields
transfer_source: bool transfer_dest: bool uniform_texel_buffer: bool storage_texel_buffer: bool uniform_buffer: bool storage_buffer: bool index_buffer: bool vertex_buffer: bool indirect_buffer: bool
Methods

impl Usage

fn none() -> Usage

Builds a Usage with all values set to false.

fn all() -> Usage

Builds a Usage with all values set to true. Can be used for quick prototyping.

fn transfer_source() -> Usage

Builds a Usage with transfer_source set to true and the rest to false.

fn transfer_dest() -> Usage

Builds a Usage with transfer_dest set to true and the rest to false.

fn vertex_buffer() -> Usage

Builds a Usage with vertex_buffer set to true and the rest to false.

fn vertex_buffer_transfer_dest() -> Usage

Builds a Usage with vertex_buffer and transfer_dest set to true and the rest to false.

fn index_buffer() -> Usage

Builds a Usage with index_buffer set to true and the rest to false.

fn index_buffer_transfer_dest() -> Usage

Builds a Usage with index_buffer and transfer_dest set to true and the rest to false.

fn uniform_buffer() -> Usage

Builds a Usage with uniform_buffer set to true and the rest to false.

fn uniform_buffer_transfer_dest() -> Usage

Builds a Usage with uniform_buffer and transfer_dest set to true and the rest to false.

fn indirect_buffer() -> Usage

Builds a Usage with indirect_buffer set to true and the rest to false.

fn indirect_buffer_transfer_dest() -> Usage

Builds a Usage with indirect_buffer and transfer_dest set to true and the rest to false.