Trait vulkano::buffer::TypedBufferAccess

pub unsafe trait TypedBufferAccess: BufferAccess {
    type Content: ?Sized;
    fn len(&self) -> usize
    where
        Self::Content: Content,
    { ... }
}

Extension trait for BufferAccess. Indicates the type of the content of the buffer.

Associated Types

type Content: ?Sized

The type of the content.

Loading content...
Provided methods

fn len(&self) -> usize where
Self::Content: Content,

Returns the length of the buffer in number of elements.

This method can only be called for buffers whose type is known to be an array.

Loading content...