Struct vulkano::buffer::traits::PipelineMemoryBarrierRequest

pub struct PipelineMemoryBarrierRequest {
    pub offset: usize,
    pub size: usize,
    pub source_access: AccessFlagBits,
    pub destination_access: AccessFlagBits,
}

Requests that a memory barrier is created as part of the pipeline barrier.

By default, a pipeline barrier only guarantees that the source operations are executed before the destination operations, but it doesn't make memory writes made by source operations visible to the destination operations. In order to make so, you have to add a memory barrier.

The memory barrier always concerns the buffer that is currently being processed. You can't add a memory barrier that concerns another resource.

Fields
offset: usize

Offset of start of the range to flush.

size: usize

Size of the range to flush.

source_access: AccessFlagBits

Source accesses.

destination_access: AccessFlagBits

Destination accesses.