Trait vulkano::buffer::traits::CommandListState

pub trait CommandListState {
    type FinishedState: CommandBufferState;
    fn transition(self, num_command: usize, buffer: &UnsafeBuffer, offset: usize, size: usize, write: bool, stage: PipelineStages, access: AccessFlagBits) -> (Self, Option<PipelineBarrierRequest>) where Self: Sized;
    fn finish(self) -> (Self::FinishedState, Option<PipelineBarrierRequest>);
}

Trait for objects that represent the state of a slice of the buffer in a list of commands.

Associated Types

type FinishedState: CommandBufferState

Required Methods

fn transition(self, num_command: usize, buffer: &UnsafeBuffer, offset: usize, size: usize, write: bool, stage: PipelineStages, access: AccessFlagBits) -> (Self, Option<PipelineBarrierRequest>) where Self: Sized

Returns a new state that corresponds to the moment after a slice of the buffer has been used in the pipeline. The parameters indicate in which way it has been used.

If the transition should result in a pipeline barrier, then it must be returned by this function.

fn finish(self) -> (Self::FinishedState, Option<PipelineBarrierRequest>)

Function called when the command buffer builder is turned into a real command buffer.

This function can return an additional pipeline barrier that will be applied at the end of the command buffer.