Struct vulkano::buffer::view::BufferView

pub struct BufferView<F, B> where B: Buffer {
    // some fields omitted
}

Represents a way for the GPU to interpret buffer data. See the documentation of the view module.

Methods

impl<F, B> BufferView<F, B> where B: Buffer

fn new<'a, S>(buffer: S, format: F) -> Result<Arc<BufferView<F, B>>, BufferViewCreationError> where S: Into<BufferSlice<'a, [F::Pixel], B>>, B: 'static, F: StrongStorage + 'static

Builds a new buffer view.

unsafe fn unchecked<'a, S, T: ?Sized>(buffer: S, format: F) -> Result<Arc<BufferView<F, B>>, BufferViewCreationError> where S: Into<BufferSlice<'a, T, B>>, B: 'static, T: 'static, F: FormatDesc + 'static

Builds a new buffer view without checking that the format is correct.

fn buffer(&self) -> &Arc<B>

Returns the buffer associated to this view.

fn uniform_texel_buffer(&self) -> bool

Returns true if the buffer view can be used as a uniform texel buffer.

fn storage_texel_buffer(&self) -> bool

Returns true if the buffer view can be used as a storage texel buffer.

fn storage_texel_buffer_atomic(&self) -> bool

Returns true if the buffer view can be used as a storage texel buffer with atomic accesses.