Struct vulkano::command_buffer::pool::StandardCommandPool

pub struct StandardCommandPool {
    // some fields omitted
}

Standard implementation of a command pool.

Will use one Vulkan pool per thread in order to avoid locking. Will try to reuse command buffers. Locking is required only when allocating/freeing command buffers.

Methods

impl StandardCommandPool

fn new(device: &Arc<Device>, queue_family: QueueFamily) -> StandardCommandPool

Builds a new pool.

Panic

Panics if the device and the queue family don't belong to the same physical device.