Trait vulkano::command_buffer::pool::CommandPoolBuilderAlloc

pub unsafe trait CommandPoolBuilderAlloc: DeviceOwned {
    type Alloc: CommandPoolAlloc;
    fn inner(&self) -> &UnsafeCommandPoolAlloc;

    fn into_alloc(self) -> Self::Alloc;

    fn queue_family(&self) -> QueueFamily;
}

A command buffer allocated from a pool and that can be recorded.

Safety

See CommandPool for information about safety.

Associated Types

type Alloc: CommandPoolAlloc

Return type of into_alloc.

Loading content...
Required methods

fn inner(&self) -> &UnsafeCommandPoolAlloc

Returns the internal object that contains the command buffer.

fn into_alloc(self) -> Self::Alloc

Turns this builder into a command buffer that is pending execution.

fn queue_family(&self) -> QueueFamily

Returns the queue family that the pool targets.

Loading content...