Module vulkano::command_buffer::std

Reexports

pub use self::empty::PrimaryCb;

pub use self::empty::PrimaryCbBuilder;

Modules

dispatch

draw

empty

fill_buffer

render_pass

update_buffer

Traits

InsideRenderPass
Extension trait for StdCommandsList that indicates that we're inside a render pass.

OutsideRenderPass
Extension trait for StdCommandsList that indicates that we're outside a render pass.

ResourcesStates
Trait for objects that hold states of buffers and images.

StdCommandsList
A list of commands that can be turned into a command buffer.