Trait vulkano::command_buffer::std::InsideRenderPass

pub unsafe trait InsideRenderPass: StdCommandsList {
    type RenderPass: RenderPass;
    type Framebuffer: Framebuffer;
    fn current_subpass(&self) -> u32;
    fn secondary_subpass(&self) -> bool;
    fn render_pass(&self) -> &Arc<Self::RenderPass>;
    fn framebuffer(&self) -> &Self::Framebuffer;
}

Extension trait for StdCommandsList that indicates that we're inside a render pass.

Associated Types

type RenderPass: RenderPass

type Framebuffer: Framebuffer

Required Methods

fn current_subpass(&self) -> u32

Returns the number of the subpass we're in. The value is 0-indexed, so immediately after calling begin_render_pass the value will be 0.

The value should always be strictly inferior to the number of subpasses in the render pass.

fn secondary_subpass(&self) -> bool

If true, only secondary command buffers can be added inside the subpass. If false, only inline draw commands can be added.

fn render_pass(&self) -> &Arc<Self::RenderPass>

fn framebuffer(&self) -> &Self::Framebuffer