Trait vulkano::command_buffer::std::ResourcesStates

pub unsafe trait ResourcesStates {
    unsafe fn extract_buffer_state<B>(&mut self, buffer: &B) -> Option<B::CommandListState> where B: TrackedBuffer;
    unsafe fn extract_image_state<I>(&mut self, image: &I) -> Option<I::CommandListState> where I: TrackedImage;
}

Trait for objects that hold states of buffers and images.

Required Methods

unsafe fn extract_buffer_state<B>(&mut self, buffer: &B) -> Option<B::CommandListState> where B: TrackedBuffer

Returns the state of a buffer, or None if the buffer hasn't been used yet.

Whether the buffer passed as parameter is the same as the one in the commands list must be determined with the is_same method of TrackedBuffer.

Calling this function extracts the state from the list, meaning that the state will be managed by the code that called this function instead of being managed by the object itself. Hence why the function is unsafe.

Panic

Panics if the state of that buffer has already been previously extracted.

unsafe fn extract_image_state<I>(&mut self, image: &I) -> Option<I::CommandListState> where I: TrackedImage

Returns the state of an image, or None if the image hasn't been used yet.

See the description of extract_buffer_state.

Panic

Panics if the state of that image has already been previously extracted.