Trait vulkano::command_buffer::std::StdCommandsList

pub unsafe trait StdCommandsList: ResourcesStates {
    type Pool: CommandPool;
    type Output: CommandBuffer<Pool=Self::Pool>;
    fn buildable_state(&self) -> bool;
    fn num_commands(&self) -> usize;
    fn check_queue_validity(&self, queue: QueueFamily) -> Result<(), ()>;
    fn is_compute_pipeline_bound<Pl>(&self, pipeline: &Arc<ComputePipeline<Pl>>) -> bool;
    fn is_graphics_pipeline_bound<Pv, Pl, Prp>(&self, pipeline: &Arc<GraphicsPipeline<Pv, Pl, Prp>>) -> bool;
    unsafe fn raw_build<I, F>(self, additional_elements: F, barriers: I, final_barrier: PipelineBarrierBuilder) -> Self::Output where F: FnOnce(&mut UnsafeCommandBufferBuilder<Self::Pool>), I: Iterator<Item=(usize, PipelineBarrierBuilder)>;

    fn update_buffer<'a, B, D: ?Sized>(self, buffer: B, data: &'a D) -> UpdateCommand<'a, Self, B, D> where Self: Sized + OutsideRenderPass, B: TrackedBuffer, D: Copy + 'static { ... }
    fn fill_buffer<B>(self, buffer: B, data: u32) -> FillCommand<Self, B> where Self: Sized + OutsideRenderPass, B: TrackedBuffer { ... }
    fn dispatch<'a, Pl, S, Pc>(self, pipeline: Arc<ComputePipeline<Pl>>, sets: S, dimensions: [u32; 3], push_constants: &'a Pc) -> DispatchCommand<'a, Self, Pl, S, Pc> where Self: Sized + StdCommandsList + OutsideRenderPass, Pl: PipelineLayout, S: TrackedDescriptorSetsCollection, Pc: 'a { ... }
    fn begin_render_pass<F, C>(self, framebuffer: F, secondary: bool, clear_values: C) -> BeginRenderPassCommand<Self, F::RenderPass, F> where Self: Sized + OutsideRenderPass, F: Framebuffer, F::RenderPass: RenderPass + RenderPassClearValues<C> { ... }
    fn next_subpass(self, secondary: bool) -> NextSubpassCommand<Self> where Self: Sized + InsideRenderPass { ... }
    fn end_render_pass(self) -> EndRenderPassCommand<Self> where Self: Sized + InsideRenderPass { ... }
    fn draw<'a, Pv, Pl, Prp, S, Pc, V>(self, pipeline: Arc<GraphicsPipeline<Pv, Pl, Prp>>, dynamic: &DynamicState, vertices: V, sets: S, push_constants: &'a Pc) -> DrawCommand<'a, Self, Pv, Pl, Prp, S, Pc> where Self: Sized + StdCommandsList + InsideRenderPass, Pl: PipelineLayout, S: TrackedDescriptorSetsCollection, Pc: 'a, Pv: Source<V> { ... }
    fn build(self) -> Self::Output where Self: Sized { ... }
}

A list of commands that can be turned into a command buffer.

Associated Types

type Pool: CommandPool

The type of the pool that will be used to create the command buffer.

type Output: CommandBuffer<Pool=Self::Pool>

The type of the command buffer that will be generated.

Required Methods

fn buildable_state(&self) -> bool

Returns true if the command buffer can be built. This function should always return true, except when we're building a primary command buffer that is inside a render pass.

fn num_commands(&self) -> usize

Returns the number of commands in the commands list.

Note that multiple actual commands may count for just 1.

fn check_queue_validity(&self, queue: QueueFamily) -> Result<(), ()>

Checks whether the command can be executed on the given queue family.

fn is_compute_pipeline_bound<Pl>(&self, pipeline: &Arc<ComputePipeline<Pl>>) -> bool

Returns true if the given compute pipeline is currently binded in the commands list.

fn is_graphics_pipeline_bound<Pv, Pl, Prp>(&self, pipeline: &Arc<GraphicsPipeline<Pv, Pl, Prp>>) -> bool

Returns true if the given graphics pipeline is currently binded in the commands list.

unsafe fn raw_build<I, F>(self, additional_elements: F, barriers: I, final_barrier: PipelineBarrierBuilder) -> Self::Output where F: FnOnce(&mut UnsafeCommandBufferBuilder<Self::Pool>), I: Iterator<Item=(usize, PipelineBarrierBuilder)>

Turns the commands list into a command buffer.

This function accepts additional arguments that will customize the output:

additional_elements is a closure that must be called on the command buffer builder after it has finished building and before final_barrier are added.

barriers is a list of pipeline barriers accompanied by a command number. The pipeline barrier must happen after the given command number. Usually you want all the the command numbers to be inferior to num_commands.

final_barrier is a pipeline barrier that must be added at the end of the command buffer builder.

This function doesn't check that buildable_state returns true.

Provided Methods

fn update_buffer<'a, B, D: ?Sized>(self, buffer: B, data: &'a D) -> UpdateCommand<'a, Self, B, D> where Self: Sized + OutsideRenderPass, B: TrackedBuffer, D: Copy + 'static

Adds a command that writes the content of a buffer.

After this command is executed, the content of buffer will become data.

fn fill_buffer<B>(self, buffer: B, data: u32) -> FillCommand<Self, B> where Self: Sized + OutsideRenderPass, B: TrackedBuffer

Adds a command that writes the content of a buffer.

fn dispatch<'a, Pl, S, Pc>(self, pipeline: Arc<ComputePipeline<Pl>>, sets: S, dimensions: [u32; 3], push_constants: &'a Pc) -> DispatchCommand<'a, Self, Pl, S, Pc> where Self: Sized + StdCommandsList + OutsideRenderPass, Pl: PipelineLayout, S: TrackedDescriptorSetsCollection, Pc: 'a

Adds a command that executes a compute shader.

The dimensions are the number of working groups to start. The GPU will execute the compute shader dimensions[0] * dimensions[1] * dimensions[2] times.

The pipeline is the compute pipeline that will be executed, and the sets and push constants will be accessible to all the invocations.

fn begin_render_pass<F, C>(self, framebuffer: F, secondary: bool, clear_values: C) -> BeginRenderPassCommand<Self, F::RenderPass, F> where Self: Sized + OutsideRenderPass, F: Framebuffer, F::RenderPass: RenderPass + RenderPassClearValues<C>

Adds a command that starts a render pass.

If secondary is true, then you will only be able to add secondary command buffers while you're inside the first subpass on the render pass. If secondary is false, you will only be able to add inline draw commands and not secondary command buffers.

You must call this before you can add draw commands.

fn next_subpass(self, secondary: bool) -> NextSubpassCommand<Self> where Self: Sized + InsideRenderPass

Adds a command that jumps to the next subpass of the current render pass.

fn end_render_pass(self) -> EndRenderPassCommand<Self> where Self: Sized + InsideRenderPass

Adds a command that ends the current render pass.

This must be called after you went through all the subpasses and before you can build the command buffer or add further commands.

fn draw<'a, Pv, Pl, Prp, S, Pc, V>(self, pipeline: Arc<GraphicsPipeline<Pv, Pl, Prp>>, dynamic: &DynamicState, vertices: V, sets: S, push_constants: &'a Pc) -> DrawCommand<'a, Self, Pv, Pl, Prp, S, Pc> where Self: Sized + StdCommandsList + InsideRenderPass, Pl: PipelineLayout, S: TrackedDescriptorSetsCollection, Pc: 'a, Pv: Source<V>

Adds a command that draws.

Can only be used from inside a render pass.

fn build(self) -> Self::Output where Self: Sized

Turns the commands list into a command buffer that can be submitted.