Struct vulkano::command_buffer::PrimaryCommandBufferBuilder

pub struct PrimaryCommandBufferBuilder {
    // some fields omitted
}

A prototype of a primary command buffer.

Usage

let commands_buffer =
    PrimaryCommandBufferBuilder::new(&device)
        .copy_memory(..., ...)
        .draw(...)
        .build();
 

Methods

impl PrimaryCommandBufferBuilder

fn new(pool: &Arc<CommandBufferPool>) -> Result<PrimaryCommandBufferBuilder, OomError>

Builds a new primary command buffer and start recording commands in it.

fn update_buffer<B, T>(self, buffer: B, data: &T) -> PrimaryCommandBufferBuilder where B: Into<BufferSlice<T>>

Writes data to a buffer.

The data is stored inside the command buffer and written to the given buffer slice. This function is intended to be used for small amounts of data (only 64kB is allowed). if you want to transfer large amounts of data, use copies between buffers.

Panic

Panicks if the size of data is not the same as the size of the buffer slice.

Panicks if the size of data is superior to 65536 bytes.

Panicks if the offset or size is not a multiple of 4.

Panicks if the buffer wasn't created with the right usage.

Panicks if the queue family doesn't support transfer operations.

unsafe fn fill_buffer<T: 'static, M>(self, buffer: &Arc<Buffer<T, M>>, offset: usize, size: usize, data: u32) -> PrimaryCommandBufferBuilder where M: MemorySourceChunk + 'static

Fills a buffer with data.

The data is repeated until it fills the range from offset to offset + size. Since the data is a u32, the offset and the size must be multiples of 4.

Panic

Panicks if offset + data is superior to the size of the buffer.

Panicks if the offset or size is not a multiple of 4.

Panicks if the buffer wasn't created with the right usage.

Panicks if the queue family doesn't support transfer operations.

Safety

Type safety is not enforced by the API.

fn copy_buffer<T: ?Sized + 'static, Ms, Md>(self, source: &Arc<Buffer<T, Ms>>, destination: &Arc<Buffer<T, Md>>) -> PrimaryCommandBufferBuilder where Ms: MemorySourceChunk + 'static, Md: MemorySourceChunk + 'static

fn copy_buffer_to_color_image<S, Ty, F, Im>(self, source: S, destination: &Arc<Image<Ty, F, Im>>) -> PrimaryCommandBufferBuilder where S: Into<BufferSlice<[F::Pixel]>>, F: StrongStorage + PossibleFloatOrCompressedFormatDesc, Ty: ImageTypeMarker

fn clear_color_image<'a, Ty, F, M>(self, image: &Arc<Image<Ty, F, M>>, color: F::ClearValue) -> PrimaryCommandBufferBuilder where Ty: ImageTypeMarker, F: PossibleFloatFormatDesc

Note that compressed formats are not supported.

fn execute_commands<'a, I>(self, iter: I) -> PrimaryCommandBufferBuilder where I: Iterator<Item=&'a SecondaryComputeCommandBuffer>

Executes secondary compute command buffers within this primary command buffer.

fn draw_inline<R, F>(self, renderpass: &Arc<RenderPass<R>>, framebuffer: &Arc<Framebuffer<F>>, clear_values: F::ClearValues) -> PrimaryCommandBufferBuilderInlineDraw where F: RenderPassLayout, R: RenderPassLayout

Start drawing on a framebuffer. This function returns an object that can be used to submit draw commands on the first subpass of the renderpass.

Panic

Panicks if the framebuffer is not compatible with the renderpass.

fn draw_secondary<R, F>(self, renderpass: &Arc<RenderPass<R>>, framebuffer: &Arc<Framebuffer<F>>, clear_values: F::ClearValues) -> PrimaryCommandBufferBuilderSecondaryDraw where F: RenderPassLayout, R: RenderPassLayout

Start drawing on a framebuffer. This function returns an object that can be used to submit secondary graphics command buffers that will operate on the first subpass of the renderpass.

Panic

Panicks if the framebuffer is not compatible with the renderpass.

fn build(self) -> Result<Arc<PrimaryCommandBuffer>, OomError>

Finish recording commands and build the command buffer.