Struct vulkano::command_buffer::PrimaryCommandBufferBuilderInlineDraw

pub struct PrimaryCommandBufferBuilderInlineDraw {
    // some fields omitted
}

Object that you obtain when calling draw_inline or next_subpass_inline.

Methods

impl PrimaryCommandBufferBuilderInlineDraw

fn draw<V, L, Pl>(self, pipeline: &Arc<GraphicsPipeline<V, Pl>>, vertices: V, dynamic: &DynamicState, sets: L) -> PrimaryCommandBufferBuilderInlineDraw where V: MultiVertex + 'static, Pl: PipelineLayoutDesc + 'static, L: DescriptorSetsCollection + 'static

Calls vkCmdDraw.

fn draw_indexed<V, L, Pl, I, Ib>(self, pipeline: &Arc<GraphicsPipeline<V, Pl>>, vertices: V, indices: Ib, dynamic: &DynamicState, sets: L) -> PrimaryCommandBufferBuilderInlineDraw where V: 'static + MultiVertex, Pl: 'static + PipelineLayoutDesc, Ib: Into<BufferSlice<[I]>>, I: 'static + Index, L: DescriptorSetsCollection + 'static

Calls vkCmdDrawIndexed.

fn next_subpass_inline(self) -> PrimaryCommandBufferBuilderInlineDraw

Switches to the next subpass of the current renderpass.

This function is similar to draw_inline on the builder.

Panic

Panicks if no more subpasses remain.

fn next_subpass_secondary(self) -> PrimaryCommandBufferBuilderSecondaryDraw

Switches to the next subpass of the current renderpass.

This function is similar to draw_secondary on the builder.

Panic

Panicks if no more subpasses remain.

fn draw_end(self) -> PrimaryCommandBufferBuilder

Finish drawing this renderpass and get back the builder.