Struct vulkano::command_buffer::PrimaryCommandBufferBuilderSecondaryDraw

pub struct PrimaryCommandBufferBuilderSecondaryDraw<P = Arc<StandardCommandPool>> where P: CommandPool {
    // some fields omitted
}

Object that you obtain when calling draw_secondary or next_subpass_secondary.

Methods

impl<P> PrimaryCommandBufferBuilderSecondaryDraw<P> where P: CommandPool

fn next_subpass_inline(self) -> PrimaryCommandBufferBuilderInlineDraw<P>

Switches to the next subpass of the current renderpass.

This function is similar to draw_inline on the builder.

Panic

Panics if no more subpasses remain.

fn next_subpass_secondary(self) -> PrimaryCommandBufferBuilderSecondaryDraw<P>

Switches to the next subpass of the current renderpass.

This function is similar to draw_secondary on the builder.

Panic

Panics if no more subpasses remain.

fn execute_commands<R, Ps>(self, cb: &Arc<SecondaryGraphicsCommandBuffer<R, Ps>>) -> PrimaryCommandBufferBuilderSecondaryDraw<P> where R: 'static + Send + Sync, Ps: CommandPool + 'static, Ps::Finished: Send + Sync + 'static

Executes secondary graphics command buffers within this primary command buffer.

Panic

Panics if the secondary command buffers wasn't created with a compatible renderpass or is using the wrong subpass.

fn draw_end(self) -> PrimaryCommandBufferBuilder<P>

Finish drawing this renderpass and get back the builder.

Panic

Panics if not at the last subpass.