Struct vulkano::command_buffer::SecondaryComputeCommandBufferBuilder

pub struct SecondaryComputeCommandBufferBuilder<P = Arc<StandardCommandPool>> where P: CommandPool {
    // some fields omitted
}

A prototype of a secondary compute command buffer.

Methods

impl SecondaryComputeCommandBufferBuilder<Arc<StandardCommandPool>>

fn new(device: &Arc<Device>, queue_family: QueueFamily) -> SecondaryComputeCommandBufferBuilder<Arc<StandardCommandPool>>

Builds a new secondary command buffer and start recording commands in it.

Panic

Panics if the device or host ran out of memory.

Panics if the device and queue family do not belong to the same physical device.

impl<P> SecondaryComputeCommandBufferBuilder<P> where P: CommandPool

fn raw(pool: P) -> Result<SecondaryComputeCommandBufferBuilder<P>, OomError>

See the docs of new().

fn update_buffer<'a, B, T, Bb>(self, buffer: B, data: &T) -> SecondaryComputeCommandBufferBuilder<P> where B: Into<BufferSlice<'a, T, Bb>>, Bb: Buffer + 'static, T: Clone + 'static + Send + Sync

Writes data to a buffer.

The data is stored inside the command buffer and written to the given buffer slice. This function is intended to be used for small amounts of data (only 64kB is allowed). if you want to transfer large amounts of data, use copies between buffers.

Panic

Panics if the size of data is not the same as the size of the buffer slice.

Panics if the size of data is superior to 65536 bytes.

Panics if the offset or size is not a multiple of 4.

Panics if the buffer wasn't created with the right usage.

Panics if the queue family doesn't support transfer operations.

unsafe fn fill_buffer<B>(self, buffer: &Arc<B>, offset: usize, size: usize, data: u32) -> SecondaryComputeCommandBufferBuilder<P> where B: Buffer + 'static

Fills a buffer with data.

The data is repeated until it fills the range from offset to offset + size. Since the data is a u32, the offset and the size must be multiples of 4.

Panic

Panics if offset + data is superior to the size of the buffer.

Panics if the offset or size is not a multiple of 4.

Panics if the buffer wasn't created with the right usage.

Panics if the queue family doesn't support transfer operations.

Safety

Type safety is not enforced by the API.

fn build_raw(self) -> Result<SecondaryComputeCommandBuffer<P>, OomError>

See the docs of build().

fn build(self) -> Arc<SecondaryComputeCommandBuffer<P>>

Finish recording commands and build the command buffer.

Panic

Panics if the device or host ran out of memory.