Struct vulkano::command_buffer::SecondaryGraphicsCommandBuffer

pub struct SecondaryGraphicsCommandBuffer<R, P = Arc<StandardCommandPool>> where P: CommandPool {
    // some fields omitted
}

Represents a collection of commands to be executed by the GPU.

A secondary graphics command buffer contains draw commands and non-draw commands. Secondary command buffers can't specify which framebuffer they are drawing to. Instead you must create a primary command buffer, specify a framebuffer, and then call the secondary command buffer.

A secondary graphics command buffer can't be called outside of a renderpass.