Struct vulkano::command_buffer::Submission

#[must_use]
pub struct Submission {
    // some fields omitted
}

Methods

impl Submission

fn destroying_would_block(&self) -> bool

Returns true if destroying this Submission object would block the CPU for some time.

fn finished(&self) -> bool

Returns true if the GPU has finished executing this submission.

fn wait(&self, timeout: Duration) -> Result<(), FenceWaitError>

Waits until the submission has finished being executed by the device.

fn queue(&self) -> &Arc<Queue>

Returns the queue the command buffers were submitted to.