Module vulkano::command_buffer::submit

Structs

Submission
Returned when you submit one or multiple command buffers.

Submit

SubmitInfo
Information about how the submitting function should synchronize the submission.

SubmitListOpaque

Traits

CommandBuffer
Trait for objects that represent commands ready to be executed by the GPU.

SubmitList