Struct vulkano::command_buffer::submit::Submission

#[must_use]
pub struct Submission {
    // some fields omitted
}

Returned when you submit one or multiple command buffers.

This object holds the resources that are used by the GPU and that must be kept alive for at least as long as the GPU is executing the submission. Therefore destroying a Submission object will block until the GPU is finished executing.

Whenever you submit a command buffer, you are encouraged to store the returned Submission in a long-living container such as a Vec. From time to time, you can clean the obsolete objects by checking whether destroying_would_block() returns false. For example, if you use a Vec you can do vec.retain(|s| s.destroying_would_block()).

Methods

impl Submission

fn destroying_would_block(&self) -> bool

Returns true if destroying this Submission object would block the CPU for some time.

fn finished(&self) -> bool

Returns true if the GPU has finished executing this submission.