Struct vulkano::command_buffer::submit::Submit

pub struct Submit<L> {
    // some fields omitted
}

Methods

impl Submit<()>

fn new() -> Submit<()>

Builds an empty submission list.

impl<L> Submit<L> where L: SubmitList

fn add<C>(self, command_buffer: C) -> Submit<(C, L)> where C: CommandBuffer + 'static

Adds a command buffer to submit to the list.

In the Vulkan API, a submission is divided into batches that each contain one or more command buffers. Vulkano will automatically determine which command buffers can be grouped into the same batch.

fn submit(self, queue: &Arc<Queue>) -> Submission

Submits the list of command buffers.