Struct vulkano::command_buffer::submit::SubmitInfo

pub struct SubmitInfo<Swi, Ssi> {
    pub semaphores_wait: Swi,
    pub semaphores_signal: Ssi,
    pub pre_pipeline_barrier: PipelineBarrierBuilder,
    pub post_pipeline_barrier: PipelineBarrierBuilder,
}

Information about how the submitting function should synchronize the submission.

Fields
semaphores_wait: Swi

List of semaphores to wait upon before the command buffer starts execution.

semaphores_signal: Ssi

List of semaphores to signal after the command buffer has finished.

pre_pipeline_barrier: PipelineBarrierBuilder

Pipeline barrier to execute on the queue and immediately before the command buffer. Ignored if empty.

post_pipeline_barrier: PipelineBarrierBuilder

Pipeline barrier to execute on the queue and immediately after the command buffer. Ignored if empty.