Module vulkano::command_buffer::sys

Lowest-level interface for command buffers creation.

This module provides the structs necessary to create command buffers. The main purpose of this module is to provide a nice API, and almost no safety checks are performed except through debug_assert!.

Safety

Each individual function is documented and indicates what exactly must be done in order to be safe.

Things to generally look for when you use an UnsafeCommandBuffer are:

The objects must be kept alive for at least as long as the command buffer.

The objects must be properly synchronized and transitionned with pipeline barriers.

The capabilities of the queue must be checked (for example running a draw operation on a transfer-only queue won't work).

Some commands can only be called from inside or outside of a render pass.

Structs

BufferCopyRegion
A copy between two buffers.

ImageSubresourcesRange
Range of an image subresource.

PipelineBarrierBuilder
Prototype for a pipeline barrier that's going to be added to a command buffer builder.

UnsafeCommandBuffer

UnsafeCommandBufferBuilder

Enums

Flags
Flags to pass when creating a command buffer.

Kind
Determines the kind of command buffer that we want to create.