Struct vulkano::command_buffer::sys::PipelineBarrierBuilder

pub struct PipelineBarrierBuilder {
    // some fields omitted
}

Prototype for a pipeline barrier that's going to be added to a command buffer builder.

Note: we use a builder-like API here so that users can pass multiple buffers or images of multiple different types. Doing so with a single function would be very tedious in terms of API.

Methods

impl PipelineBarrierBuilder

fn new() -> PipelineBarrierBuilder

Adds a command that adds a pipeline barrier to a command buffer.

fn is_empty(&self) -> bool

Returns true if no barrier or execution dependency has been added yet.

fn merge(&mut self, other: PipelineBarrierBuilder)

Merges another pipeline builder into this one.

unsafe fn add_execution_dependency(&mut self, source: PipelineStages, dest: PipelineStages, by_region: bool)

Adds an execution dependency. This means that all the stages in source of the previous commands must finish before any of the stages in dest of the following commands can start.

Safety

If the pipeline stages include geometry or tessellation stages, then the corresponding features must have been enabled.

There are certain rules regarding the pipeline barriers inside render passes.

unsafe fn add_memory_barrier(&mut self, source_stage: PipelineStages, source_access: AccessFlagBits, dest_stage: PipelineStages, dest_access: AccessFlagBits, by_region: bool)

Adds a memory barrier. This means that all the memory writes by the given source stages for the given source accesses must be visible by the given dest stages for the given dest accesses.

Also adds an execution dependency.

Safety

If the pipeline stages include geometry or tessellation stages, then the corresponding features must have been enabled.

There are certain rules regarding the pipeline barriers inside render passes.

unsafe fn add_buffer_barrier_request(&mut self, buffer: &UnsafeBuffer, request: BufferPipelineBarrierRequest)

unsafe fn add_image_barrier_request(&mut self, image: &UnsafeImage, request: ImagePipelineBarrierRequest)

unsafe fn add_buffer_memory_barrier<'a, T: ?Sized, B>(&mut self, buffer: BufferSlice<'a, T, B>, source_stage: PipelineStages, source_access: AccessFlagBits, dest_stage: PipelineStages, dest_access: AccessFlagBits, by_region: bool, queue_transfer: Option<(u32, u32)>) where B: Buffer

Adds a buffer memory barrier. This means that all the memory writes to the given buffer by the given source stages for the given source accesses must be visible by the given dest stages for the given dest accesses.

Also adds an execution dependency.

Also allows transfering buffer ownership between queues.

Safety

If the pipeline stages include geometry or tessellation stages, then the corresponding features must have been enabled.

There are certain rules regarding the pipeline barriers inside render passes.

The buffer must be alive for at least as long as the command buffer to which this barrier is added.

Queue ownership transfers must be correct.

unsafe fn add_image_memory_barrier<I>(&mut self, image: &Arc<I>, mipmaps: Range<u32>, layers: Range<u32>, source_stage: PipelineStages, source_access: AccessFlagBits, dest_stage: PipelineStages, dest_access: AccessFlagBits, by_region: bool, queue_transfer: Option<(u32, u32)>, current_layout: Layout, new_layout: Layout) where I: Image

Adds an image memory barrier. This is the equivalent of add_buffer_memory_barrier but for images.

In addition to transfering image ownership between queues, it also allows changing the layout of images.

Safety

If the pipeline stages include geometry or tessellation stages, then the corresponding features must have been enabled.

There are certain rules regarding the pipeline barriers inside render passes.

The buffer must be alive for at least as long as the command buffer to which this barrier is added.

Queue ownership transfers must be correct.

Image layouts transfers must be correct.

Access flags must be compatible with the image usage flags passed at image creation.