Module vulkano::command_buffer::validity

Functions that check the validity of commands.

Structs

CheckCopyBuffer
Information returned if check_copy_buffer succeeds.

CheckIndexBuffer
Information returned if check_index_buffer succeeds.

CheckVertexBuffer
Information returned if check_vertex_buffer succeeds.

Enums

CheckBlitImageError
Error that can happen from check_clear_color_image.

CheckClearColorImageError
Error that can happen from check_clear_color_image.

CheckCopyBufferError
Error that can happen from check_copy_buffer.

CheckCopyBufferImageError
Error that can happen from check_copy_buffer_image.

CheckCopyBufferImageTy
Type of operation to check.

CheckCopyImageError
Error that can happen from check_copy_image.

CheckDescriptorSetsValidityError
Error that can happen when checking descriptor sets validity.

CheckDispatchError
Error that can happen when checking dispatch command validity.

CheckDynamicStateValidityError
Error that can happen when validating dynamic states.

CheckFillBufferError
Error that can happen when attempting to add a fill_buffer command.

CheckIndexBufferError
Error that can happen when checking whether binding an index buffer is valid.

CheckPushConstantsValidityError
Error that can happen when checking push constants validity.

CheckUpdateBufferError
Error that can happen when attempting to add an update_buffer command.

CheckVertexBufferError
Error that can happen when checking whether the vertex buffers are valid.

Functions

check_blit_image
Checks whether a blit image command is valid.

check_clear_color_image
Checks whether a clear color image command is valid.

check_copy_buffer
Checks whether a copy buffer command is valid.

check_copy_buffer_image
Checks whether a copy buffer-image command is valid. Can check both buffer-to-image copies and image-to-buffer copies.

check_copy_image
Checks whether a copy image command is valid.

check_descriptor_sets_validity
Checks whether descriptor sets are compatible with the pipeline.

check_dispatch
Checks whether the dispatch dimensions are supported by the device.

check_dynamic_state_validity
Checks whether states that are about to be set are correct.

check_fill_buffer
Checks whether a fill buffer command is valid.

check_index_buffer
Checks whether an index buffer can be bound.

check_push_constants_validity
Checks whether push constants are compatible with the pipeline.

check_update_buffer
Checks whether an update buffer command is valid.

check_vertex_buffers
Checks whether vertex buffers can be bound.