Enum vulkano::command_buffer::validity::CheckBlitImageError

pub enum CheckBlitImageError {
    MissingTransferSourceUsage,
    MissingTransferDestinationUsage,
    SourceFormatNotSupported,
    DestinationFormatNotSupported,
    DepthStencilNearestMandatory,
    DepthStencilFormatMismatch,
    IncompatibleFormatsTypes {
        source_format_ty: FormatTy,
        destination_format_ty: FormatTy,
    },
    UnexpectedMultisampled,
    SourceCoordinatesOutOfRange,
    DestinationCoordinatesOutOfRange,
    IncompatibleRangeForImageType,
}

Error that can happen from check_clear_color_image.

Variants
MissingTransferSourceUsage

The source is missing the transfer source usage.

MissingTransferDestinationUsage

The destination is missing the transfer destination usage.

SourceFormatNotSupported

The format of the source image doesn't support blit operations.

DestinationFormatNotSupported

The format of the destination image doesn't support blit operations.

DepthStencilNearestMandatory

You must use the nearest filter when blitting depth/stencil images.

DepthStencilFormatMismatch

The format of the source and destination must be equal when blitting depth/stencil images.

IncompatibleFormatsTypes

The types of the source format and the destination format aren't compatible.

Fields of IncompatibleFormatsTypes

source_format_ty: FormatTydestination_format_ty: FormatTy
UnexpectedMultisampled

Blitting between multisampled images is forbidden.

SourceCoordinatesOutOfRange

The offsets, array layers and/or mipmap levels are out of range in the source image.

DestinationCoordinatesOutOfRange

The offsets, array layers and/or mipmap levels are out of range in the destination image.

IncompatibleRangeForImageType

The top-left and/or bottom-right coordinates are incompatible with the image type.