Enum vulkano::descriptor::descriptor::DescriptorDescSupersetError

pub enum DescriptorDescSupersetError {
    ArrayTooSmall {
        len: u32,
        required: u32,
    },
    TypeMismatch,
    MutabilityRequired,
    ShaderStagesNotSuperset,
    DimensionsMismatch {
        provided: DescriptorImageDescDimensions,
        expected: DescriptorImageDescDimensions,
    },
    FormatMismatch {
        provided: Option<Format>,
        expected: Format,
    },
    MultisampledMismatch {
        provided: bool,
        expected: bool,
    },
    IncompatibleArrayLayers {
        provided: DescriptorImageDescArray,
        required: DescriptorImageDescArray,
    },
}

Error when checking whether a descriptor is a superset of another one.

Variants

ArrayTooSmall

The number of array elements of the descriptor is smaller than expected.

Fields of ArrayTooSmall

len: u32required: u32

TypeMismatch

The descriptor type doesn't match the type of the other descriptor.

MutabilityRequired

The descriptor is marked as read-only, but the other is not.

ShaderStagesNotSuperset

The shader stages are not a superset of one another.

DimensionsMismatch

Fields of DimensionsMismatch

provided: DescriptorImageDescDimensionsexpected: DescriptorImageDescDimensions

FormatMismatch

Fields of FormatMismatch

provided: Option<Format>expected: Format

MultisampledMismatch

Fields of MultisampledMismatch

provided: boolexpected: bool

IncompatibleArrayLayers

Fields of IncompatibleArrayLayers

provided: DescriptorImageDescArrayrequired: DescriptorImageDescArray