Trait vulkano::descriptor::descriptor_set::collection::DescriptorSetsCollection

pub unsafe trait DescriptorSetsCollection {
    fn into_vec(self) -> Vec<Box<dyn DescriptorSet + Send + Sync>>;

    fn num_bindings_in_set(&self, set: usize) -> Option<usize>;

    fn descriptor(&self, set: usize, binding: usize) -> Option<DescriptorDesc>;
}

A collection of descriptor set objects.

Required methods

fn into_vec(self) -> Vec<Box<dyn DescriptorSet + Send + Sync>>

fn num_bindings_in_set(&self, set: usize) -> Option<usize>

Returns the number of descriptors in the set. Includes possibly empty descriptors.

Returns None if the set is out of range.

fn descriptor(&self, set: usize, binding: usize) -> Option<DescriptorDesc>

Returns the descriptor for the given binding of the given set.

Returns None if out of range.

Loading content...