Struct vulkano::descriptor::descriptor_set::DescriptorPool

pub struct DescriptorPool {
    // some fields omitted
}

Pool from which descriptor sets are allocated from.

A pool has a maximum number of descriptor sets and a maximum number of descriptors (one value per descriptor type) it can allocate.

Methods

impl DescriptorPool

fn raw(device: &Arc<Device>) -> Result<DescriptorPool, OomError>

See the docs of new().

fn new(device: &Arc<Device>) -> Arc<DescriptorPool>

Initializes a new pool.

Panic

Panics if the device or host ran out of memory.

fn device(&self) -> &Arc<Device>

Returns the device this pool was created from.