Struct vulkano::descriptor::descriptor_set::FixedSizeDescriptorSetBuilderArray

pub struct FixedSizeDescriptorSetBuilderArray<'a, L: 'a, R> { /* fields omitted */ }

Same as FixedSizeDescriptorSetBuilder, but we're in an array.

Implementations

impl<'a, L, R> FixedSizeDescriptorSetBuilderArray<'a, L, R> where
L: PipelineLayoutAbstract,

pub fn leave_array(
self
) -> Result<FixedSizeDescriptorSetBuilder<'a, L, R>, PersistentDescriptorSetError>

Leaves the array. Call this once you added all the elements of the array.

pub fn add_buffer<T>(
self,
buffer: T
) -> Result<FixedSizeDescriptorSetBuilderArray<'a, L, (R, PersistentDescriptorSetBuf<T>)>, PersistentDescriptorSetError> where
T: BufferAccess,

Binds a buffer as the next element in the array.

An error is returned if the buffer isn't compatible with the descriptor.

Panic

Panics if the buffer doesn't have the same device as the pipeline layout.

pub fn add_buffer_view<T>(
self,
view: T
) -> Result<FixedSizeDescriptorSetBuilderArray<'a, L, (R, PersistentDescriptorSetBufView<T>)>, PersistentDescriptorSetError> where
T: BufferViewRef,

Binds a buffer view as the next element in the array.

An error is returned if the buffer isn't compatible with the descriptor.

Panic

Panics if the buffer view doesn't have the same device as the pipeline layout.

pub fn add_image<T>(
self,
image_view: T
) -> Result<FixedSizeDescriptorSetBuilderArray<'a, L, (R, PersistentDescriptorSetImg<T>)>, PersistentDescriptorSetError> where
T: ImageViewAccess,

Binds an image view as the next element in the array.

An error is returned if the image view isn't compatible with the descriptor.

Panic

Panics if the image view doesn't have the same device as the pipeline layout.

pub fn add_sampled_image<T>(
self,
image_view: T,
sampler: Arc<Sampler>
) -> Result<FixedSizeDescriptorSetBuilderArray<'a, L, ((R, PersistentDescriptorSetImg<T>), PersistentDescriptorSetSampler)>, PersistentDescriptorSetError> where
T: ImageViewAccess,

Binds an image view with a sampler as the next element in the array.

An error is returned if the image view isn't compatible with the descriptor.

Panic

Panics if the image or the sampler doesn't have the same device as the pipeline layout.

pub fn add_sampler(
self,
sampler: Arc<Sampler>
) -> Result<FixedSizeDescriptorSetBuilderArray<'a, L, (R, PersistentDescriptorSetSampler)>, PersistentDescriptorSetError>

Binds a sampler as the next element in the array.

An error is returned if the sampler isn't compatible with the descriptor.

Panic

Panics if the sampler doesn't have the same device as the pipeline layout.