Struct vulkano::descriptor::descriptor_set::UnsafeDescriptorSet

pub struct UnsafeDescriptorSet { /* fields omitted */ }

Low-level descriptor set.

Contrary to most other objects in this library, this one doesn't free itself automatically and doesn't hold the pool or the device it is associated to. Instead it is an object meant to be used with the UnsafeDescriptorPool.

Methods

impl UnsafeDescriptorSet

pub unsafe fn write<I>(&mut self, device: &Device, writes: I) where
I: Iterator<Item = DescriptorWrite>,

Modifies a descriptor set. Doesn't check that the writes or copies are correct, and doesn't check whether the descriptor set is in use.

Important: You must ensure that the UnsafeDescriptorSetLayout object is alive before updating a descriptor set.

Safety

The Device must be the device the pool of this set was created with.

The UnsafeDescriptorSetLayout object this set was created with must be alive.

Doesn't verify that the things you write in the descriptor set match its layout.

Doesn't keep the resources alive. You have to do that yourself.

Updating a descriptor set obeys synchronization rules that aren't checked here. Once a command buffer contains a pointer/reference to a descriptor set, it is illegal to write to it.