Struct vulkano::descriptor::descriptor_set::UnsafeDescriptorSetLayout

pub struct UnsafeDescriptorSetLayout {
    // some fields omitted
}

Describes to the Vulkan implementation the layout of all descriptors within a descriptor set.

Despite its name, this type is technically not unsafe. However it serves the same purpose in the API as other types whose names start with Unsafe.

Methods

impl UnsafeDescriptorSetLayout

fn raw<I>(device: &Arc<Device>, descriptors: I) -> Result<UnsafeDescriptorSetLayout, OomError> where I: IntoIterator<Item=DescriptorDesc>

See the docs of new().

fn new<I>(device: &Arc<Device>, descriptors: I) -> Arc<UnsafeDescriptorSetLayout> where I: IntoIterator<Item=DescriptorDesc>

Builds a new UnsafeDescriptorSetLayout with the given descriptors.

Panic

Panicks if the device or host ran out of memory.

fn device(&self) -> &Arc<Device>

Returns the device used to create this layout.