Trait vulkano::descriptor::descriptor_set::DescriptorPool

pub unsafe trait DescriptorPool: DeviceOwned {
    type Alloc: DescriptorPoolAlloc;
    fn alloc(
        &mut self, 
        layout: &UnsafeDescriptorSetLayout
    ) -> Result<Self::Alloc, OomError>;
}

A pool from which descriptor sets can be allocated.

Since the destructor of Alloc must free the descriptor set, this trait is usually implemented on Arc<T> or &'a T and not T directly, so that the Alloc object can hold the pool.

Associated Types

type Alloc: DescriptorPoolAlloc

Object that represented an allocated descriptor set.

The destructor of this object should free the descriptor set.

Loading content...
Required methods

fn alloc(
&mut self,
layout: &UnsafeDescriptorSetLayout
) -> Result<Self::Alloc, OomError>

Allocates a descriptor set.

Loading content...