Trait vulkano::descriptor::descriptor_set::DescriptorSetDesc

pub unsafe trait DescriptorSetDesc {
    fn num_bindings(&self) -> usize;

    fn descriptor(&self, binding: usize) -> Option<DescriptorDesc>;
}

Trait for objects that describe the layout of the descriptors of a set.

Required methods

fn num_bindings(&self) -> usize

Returns the number of binding slots in the set.

fn descriptor(&self, binding: usize) -> Option<DescriptorDesc>

Returns a description of a descriptor, or None if out of range.

Loading content...