Trait vulkano::descriptor::descriptor_set::TrackedDescriptorSet

pub unsafe trait TrackedDescriptorSet: DescriptorSet {
    type State: TrackedDescriptorSetState<Finished=Self::Finished>;
    type Finished: TrackedDescriptorSetFinished;
    unsafe fn extract_states_and_transition<L>(&self, list: &mut L) -> (Self::State, usize, PipelineBarrierBuilder) where L: ResourcesStates;
}

Extension trait for descriptor sets so that it can be used with the standard commands list interface.

Associated Types

type State: TrackedDescriptorSetState<Finished=Self::Finished>

type Finished: TrackedDescriptorSetFinished

Required Methods

unsafe fn extract_states_and_transition<L>(&self, list: &mut L) -> (Self::State, usize, PipelineBarrierBuilder) where L: ResourcesStates

Extracts the states relevant to the buffers and images contained in the descriptor set. Then transitions them to the right state.