Trait vulkano::descriptor::descriptor_set::TrackedDescriptorSetFinished

pub unsafe trait TrackedDescriptorSetFinished {
    type SemaphoresWaitIterator: Iterator<Item=(Arc<Semaphore>, PipelineStages)>;
    type SemaphoresSignalIterator: Iterator<Item=Arc<Semaphore>>;
    unsafe fn on_submit<F>(&self, queue: &Arc<Queue>, fence: F) -> SubmitInfo<Self::SemaphoresWaitIterator, Self::SemaphoresSignalIterator> where F: FnMut() -> Arc<Fence>;
}

Associated Types

type SemaphoresWaitIterator: Iterator<Item=(Arc<Semaphore>, PipelineStages)>

Iterator that returns the list of semaphores to wait upon before the command buffer is submitted.

type SemaphoresSignalIterator: Iterator<Item=Arc<Semaphore>>

Iterator that returns the list of semaphores to signal after the command buffer has finished execution.

Required Methods

unsafe fn on_submit<F>(&self, queue: &Arc<Queue>, fence: F) -> SubmitInfo<Self::SemaphoresWaitIterator, Self::SemaphoresSignalIterator> where F: FnMut() -> Arc<Fence>