Trait vulkano::descriptor::descriptor_set::TrackedDescriptorSetState

pub unsafe trait TrackedDescriptorSetState: ResourcesStates {
    type Finished: TrackedDescriptorSetFinished;
    unsafe fn extract_buffer_state<B>(&mut self, buffer: &B) -> Option<B::CommandListState> where B: TrackedBuffer;
    unsafe fn extract_image_state<I>(&mut self, image: &I) -> Option<I::CommandListState> where I: TrackedImage;
    unsafe fn finish(self) -> (Self::Finished, PipelineBarrierBuilder);
}

Associated Types

type Finished: TrackedDescriptorSetFinished

Required Methods

unsafe fn extract_buffer_state<B>(&mut self, buffer: &B) -> Option<B::CommandListState> where B: TrackedBuffer

Extracts the state of a buffer of the descriptor set, or None if the buffer isn't in the descriptor set.

Whether the buffer passed as parameter is the same as the one in the descriptor set must be determined with the is_same method of TrackedBuffer.

Panic

Panics if the state of that buffer has already been previously extracted.

unsafe fn extract_image_state<I>(&mut self, image: &I) -> Option<I::CommandListState> where I: TrackedImage

Returns the state of an image, or None if the image isn't in the descriptor set.

See the description of extract_buffer_state.

Panic

Panics if the state of that image has already been previously extracted.

unsafe fn finish(self) -> (Self::Finished, PipelineBarrierBuilder)

Turns the object into a TrackedDescriptorSetFinished. All the buffers and images whose state hasn't been extracted must be have finished() called on them as well.

The function returns a pipeline barrier to append at the end of the command buffer.