Enum vulkano::descriptor::pipeline_layout::PipelineLayoutLimitsError

pub enum PipelineLayoutLimitsError {
    MaxDescriptorSetsLimitExceeded {
        limit: usize,
        requested: usize,
    },
    MaxPushConstantsSizeExceeded {
        limit: usize,
        requested: usize,
    },
    MaxPerStageResourcesLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxPerStageDescriptorSamplersLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxPerStageDescriptorUniformBuffersLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxPerStageDescriptorStorageBuffersLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxPerStageDescriptorSampledImagesLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxPerStageDescriptorStorageImagesLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxPerStageDescriptorInputAttachmentsLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxDescriptorSetSamplersLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxDescriptorSetUniformBuffersLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxDescriptorSetUniformBuffersDynamicLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxDescriptorSetStorageBuffersLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxDescriptorSetStorageBuffersDynamicLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxDescriptorSetSampledImagesLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxDescriptorSetStorageImagesLimitExceeded {
        limit: u32,
        requested: u32,
    },
    MaxDescriptorSetInputAttachmentsLimitExceeded {
        limit: u32,
        requested: u32,
    },
}

The pipeline layout description isn't compatible with the hardware limits.

Variants

MaxDescriptorSetsLimitExceeded

The maximum number of descriptor sets has been exceeded.

Fields of MaxDescriptorSetsLimitExceeded

limit: usize

The limit that must be fulfilled.

requested: usize

What was requested.

MaxPushConstantsSizeExceeded

The maximum size of push constants has been exceeded.

Fields of MaxPushConstantsSizeExceeded

limit: usize

The limit that must be fulfilled.

requested: usize

What was requested.

MaxPerStageResourcesLimitExceeded

The max_per_stage_resources() limit has been exceeded.

Fields of MaxPerStageResourcesLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxPerStageDescriptorSamplersLimitExceeded

The max_per_stage_descriptor_samplers() limit has been exceeded.

Fields of MaxPerStageDescriptorSamplersLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxPerStageDescriptorUniformBuffersLimitExceeded

The max_per_stage_descriptor_uniform_buffers() limit has been exceeded.

Fields of MaxPerStageDescriptorUniformBuffersLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxPerStageDescriptorStorageBuffersLimitExceeded

The max_per_stage_descriptor_storage_buffers() limit has been exceeded.

Fields of MaxPerStageDescriptorStorageBuffersLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxPerStageDescriptorSampledImagesLimitExceeded

The max_per_stage_descriptor_sampled_images() limit has been exceeded.

Fields of MaxPerStageDescriptorSampledImagesLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxPerStageDescriptorStorageImagesLimitExceeded

The max_per_stage_descriptor_storage_images() limit has been exceeded.

Fields of MaxPerStageDescriptorStorageImagesLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxPerStageDescriptorInputAttachmentsLimitExceeded

The max_per_stage_descriptor_input_attachments() limit has been exceeded.

Fields of MaxPerStageDescriptorInputAttachmentsLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxDescriptorSetSamplersLimitExceeded

The max_descriptor_set_samplers() limit has been exceeded.

Fields of MaxDescriptorSetSamplersLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxDescriptorSetUniformBuffersLimitExceeded

The max_descriptor_set_uniform_buffers() limit has been exceeded.

Fields of MaxDescriptorSetUniformBuffersLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxDescriptorSetUniformBuffersDynamicLimitExceeded

The max_descriptor_set_uniform_buffers_dynamic() limit has been exceeded.

Fields of MaxDescriptorSetUniformBuffersDynamicLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxDescriptorSetStorageBuffersLimitExceeded

The max_descriptor_set_storage_buffers() limit has been exceeded.

Fields of MaxDescriptorSetStorageBuffersLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxDescriptorSetStorageBuffersDynamicLimitExceeded

The max_descriptor_set_storage_buffers_dynamic() limit has been exceeded.

Fields of MaxDescriptorSetStorageBuffersDynamicLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxDescriptorSetSampledImagesLimitExceeded

The max_descriptor_set_sampled_images() limit has been exceeded.

Fields of MaxDescriptorSetSampledImagesLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxDescriptorSetStorageImagesLimitExceeded

The max_descriptor_set_storage_images() limit has been exceeded.

Fields of MaxDescriptorSetStorageImagesLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.

MaxDescriptorSetInputAttachmentsLimitExceeded

The max_descriptor_set_input_attachments() limit has been exceeded.

Fields of MaxDescriptorSetInputAttachmentsLimitExceeded

limit: u32

The limit that must be fulfilled.

requested: u32

What was requested.