Struct vulkano::descriptor::pipeline_layout::UnsafePipelineLayout

pub struct UnsafePipelineLayout {
    // some fields omitted
}

Low-level struct that represents the layout of the resources available to your shaders.

Despite its name, this type is technically not unsafe. However it serves the same purpose in the API as other types whose names start with Unsafe.

Methods

impl UnsafePipelineLayout

fn new<'a, I, P>(device: &Arc<Device>, layouts: I, push_constants: P) -> Result<UnsafePipelineLayout, UnsafePipelineLayoutCreationError> where I: IntoIterator<Item=&'a Arc<UnsafeDescriptorSetLayout>>, P: IntoIterator<Item=(usize, usize, ShaderStages)>

Creates a new UnsafePipelineLayout.

Panic

Panics if one of the UnsafeDescriptorSetLayout was not created with device.

fn descriptor_set_layout(&self, index: usize) -> Option<&Arc<UnsafeDescriptorSetLayout>>

Returns the UnsafeDescriptorSetLayout object of the specified set index.

Returns None if out of range.

fn device(&self) -> &Arc<Device>

Returns the device used to create this pipeline layout.