Trait vulkano::descriptor::pipeline_layout::PipelineLayoutSuperset

pub unsafe trait PipelineLayoutSuperset<Other: ?Sized>: PipelineLayoutDesc where
    Other: PipelineLayoutDesc, {
    fn ensure_superset_of(
        &self, 
        _: &Other
    ) -> Result<(), PipelineLayoutNotSupersetError>;
}

Traits that allow determining whether a pipeline layout is a superset of another one.

This trait is automatically implemented on all types that implement PipelineLayoutAbstract. TODO: once specialization lands, we can add implementations that don't perform deep comparisons

Required methods

fn ensure_superset_of(
&self,
_: &Other
) -> Result<(), PipelineLayoutNotSupersetError>

Makes sure that self is a superset of Other. Returns an Err if this is not the case.

Loading content...