Trait vulkano::descriptor::PipelineLayoutAbstract

pub unsafe trait PipelineLayoutAbstract: PipelineLayoutDesc + DeviceOwned {
    fn sys(&self) -> PipelineLayoutSys;

    fn descriptor_set_layout(
        &self, 
        index: usize
    ) -> Option<&Arc<UnsafeDescriptorSetLayout>>;
}

Trait for objects that describe the layout of the descriptors and push constants of a pipeline.

Required methods

fn sys(&self) -> PipelineLayoutSys

Returns an opaque object that allows internal access to the pipeline layout.

Can be obtained by calling PipelineLayoutAbstract::sys() on the pipeline layout.

Note: This is an internal function that you normally don't need to call.

fn descriptor_set_layout(
&self,
index: usize
) -> Option<&Arc<UnsafeDescriptorSetLayout>>

Returns the UnsafeDescriptorSetLayout object of the specified set index.

Returns None if out of range or if the set is empty for this index.

Loading content...