Module vulkano::descriptor_set

Collection of resources accessed by the pipeline.

The resources accessed by the pipeline must be accessed through what is called a descriptor. Descriptors are grouped in what is called descriptor sets. Descriptor sets are also grouped in what is called a pipeline layout.

Pipeline initialization

In order to build a pipeline object (a GraphicsPipeline or a ComputePipeline), you have to pass a pointer to a PipelineLayout<T> struct. This struct is a wrapper around a Vulkan struct that contains all the data about the descriptor sets and descriptors that will be available in the pipeline. The T parameter must implement the PipelineLayoutDesc trait and describes the descriptor sets and descriptors on vulkano's side.

To build a PipelineLayout, you need to pass a collection of DescriptorSetLayout structs. A DescriptorSetLayout<T> if the equivalent of PipelineLayout but for a single descriptor set. The T parameter must implement the DescriptorSetDesc trait.

Binding resources

In parallel of the pipeline initialization, you have to create a DescriptorSet<T>. This struct contains the list of actual resources that will be bound when the pipeline is executed. To build a DescriptorSet<T>, you need to pass a DescriptorSetLayout<T>. The T parameter must implement DescriptorSetDesc as if the same for both the descriptor set and its layout.

TODO: describe descriptor set writes

Shader analyser

While you can manually implement the PipelineLayoutDesc and DescriptorSetDesc traits on your own types, it is encouraged to use the vulkano-shaders crate instead. This crate will automatically parse your SPIR-V code and generate structs that implement these traits and describe the pipeline layout to vulkano.

Structs

DescriptorDesc
Describes a single descriptor.

DescriptorPool
Pool from which descriptor sets are allocated from.

DescriptorSet
An actual descriptor set with the resources that are binded to it.

DescriptorSetLayout
Describes the layout of all descriptors within a descriptor set.

DescriptorWrite

EmptyPipelineDesc
Implementation of PipelineLayoutDesc that describes an empty pipeline.

PipelineLayout
A collection of DescriptorSetLayout structs.

RuntimeDesc
FIXME: it should be unsafe to create this struct

RuntimeDescriptorSetDesc
FIXME: should be unsafe to create this struct

ShaderStages
Describes which shader stages have access to a descriptor.

Enums

DescriptorBind

DescriptorType
Describes what kind of resource may later be bound to a descriptor.

Traits

AbstractDescriptorSet
Implemented on all DescriptorSet objects. Hides the template parameters.

AbstractDescriptorSetLayout
Implemented on all DescriptorSetLayout objects. Hides the template parameters.

DescriptorSetDesc
Types that describe a single descriptor set.

DescriptorSetsCollection
A collection of descriptor set objects.

PipelineLayoutDesc
Types that describe the layout of a pipeline (descriptor sets and push constants).