Struct vulkano::descriptor_set::DescriptorSet

pub struct DescriptorSet<S> {
    // some fields omitted
}

An actual descriptor set with the resources that are binded to it.

Methods

impl<S> DescriptorSet<S> where S: DescriptorSetDesc

fn new(pool: &Arc<DescriptorPool>, layout: &Arc<DescriptorSetLayout<S>>, init: S::Init) -> Result<Arc<DescriptorSet<S>>, OomError>

Panic

Panicks if the pool and the layout were not created from the same Device.

unsafe fn uninitialized(pool: &Arc<DescriptorPool>, layout: &Arc<DescriptorSetLayout<S>>) -> Result<Arc<DescriptorSet<S>>, OomError>

Panic

Panicks if the pool and the layout were not created from the same Device.

fn write(&mut self, write: S::Write)

Modifies a descriptor set.

The parameter depends on your implementation of DescriptorSetDesc.

This function trusts the implementation of DescriptorSetDesc when it comes to making sure that the correct resource type is written to the correct descriptor.

unsafe fn unchecked_write(&mut self, write: Vec<DescriptorWrite>)

Modifies a descriptor set without checking that the writes are correct.