Trait vulkano::descriptor_set::DescriptorSetDesc

pub unsafe trait DescriptorSetDesc {
    type Write;
    type Init;
    fn descriptors(&self) -> Vec<DescriptorDesc>;
    fn decode_write(&self, Self::Write) -> Vec<DescriptorWrite>;
    fn decode_init(&self, Self::Init) -> Vec<DescriptorWrite>;

    fn is_compatible_with<S>(&self, _: &S) -> bool where S: DescriptorSetDesc { ... }
}

Types that describe a single descriptor set.

Associated Types

type Write

Represents a modification of a descriptor set. A parameter of this type must be passed when you modify a descriptor set.

type Init

Contains the list of attachments that must be passed at the initialization of a descriptor set object.

Required Methods

fn descriptors(&self) -> Vec<DescriptorDesc>

Returns the list of descriptors contained in this set.

fn decode_write(&self, Self::Write) -> Vec<DescriptorWrite>

Turns the Write associated type into something vulkano can understand.

fn decode_init(&self, Self::Init) -> Vec<DescriptorWrite>

Turns the Init associated type into something vulkano can understand.

Provided Methods

fn is_compatible_with<S>(&self, _: &S) -> bool where S: DescriptorSetDesc