Trait vulkano::descriptor_set::DescriptorSetsCollection

pub unsafe trait DescriptorSetsCollection {
    type Iter: ExactSizeIterator<Item=Arc<AbstractDescriptorSet>>;
    fn list(&self) -> Self::Iter;
    fn is_compatible_with<P>(&self, pipeline_layout: &Arc<PipelineLayout<P>>) -> bool;
}

A collection of descriptor set objects.

Associated Types

type Iter: ExactSizeIterator<Item=Arc<AbstractDescriptorSet>>

An iterator that produces the list of descriptor set objects contained in this collection.

Required Methods

fn list(&self) -> Self::Iter

Returns the list of descriptor set objects of this collection.

fn is_compatible_with<P>(&self, pipeline_layout: &Arc<PipelineLayout<P>>) -> bool