Trait vulkano::descriptor_set::PipelineLayoutDesc

pub unsafe trait PipelineLayoutDesc {
    type DescriptorSets;
    type DescriptorSetLayouts;
    type PushConstants;
    fn decode_descriptor_sets(&self, Self::DescriptorSets) -> Vec<Arc<AbstractDescriptorSet>>;
    fn decode_descriptor_set_layouts(&self, Self::DescriptorSetLayouts) -> Vec<Arc<AbstractDescriptorSetLayout>>;

    fn is_compatible_with<P>(&self, _: &P) -> bool where P: PipelineLayoutDesc { ... }
}

Types that describe the layout of a pipeline (descriptor sets and push constants).

Associated Types

type DescriptorSets

Represents a collection of DescriptorSet structs. A parameter of this type must be passed when you add a draw command to a command buffer that uses this layout.

type DescriptorSetLayouts

Represents a collection of DescriptorSetLayout structs. A parameter of this type must be passed when creating a PipelineLayout struct.

type PushConstants

Not yet implemented. Useless for now.

Required Methods

fn decode_descriptor_sets(&self, Self::DescriptorSets) -> Vec<Arc<AbstractDescriptorSet>>

Turns the DescriptorSets associated type into something vulkano can understand.

fn decode_descriptor_set_layouts(&self, Self::DescriptorSetLayouts) -> Vec<Arc<AbstractDescriptorSetLayout>>

Turns the DescriptorSetLayouts associated type into something vulkano can understand.

Provided Methods

fn is_compatible_with<P>(&self, _: &P) -> bool where P: PipelineLayoutDesc