Struct vulkano::device::Queue

pub struct Queue {
    // some fields omitted
}

Represents a queue where commands can be submitted.

Methods

impl Queue

fn device(&self) -> &Arc<Device>

Returns the device this queue belongs to.

fn family(&self) -> QueueFamily

Returns the family this queue belongs to.

fn id_within_family(&self) -> u32

Returns the index of this queue within its family.

fn wait_raw(&self) -> Result<(), OomError>

See the docs of wait().

fn wait(&self)

Waits until all work on this queue has finished.

Just like Device::wait(), you shouldn't have to call this function.

Panic

Panics if the device or host ran out of memory.