Trait vulkano::format::PossibleFloatFormatDesc

pub unsafe trait PossibleFloatFormatDesc: FormatDesc {
    fn is_float(&self) -> bool;
}

Trait for types that can possibly describe a float attachment.

Required methods

fn is_float(&self) -> bool

Returns true if the format is a float format.

Loading content...