Module vulkano::framebuffer

Targets on which your draw commands are executed.

There are two concepts in Vulkan:

A RenderPass is a collection of rendering passes (called subpasses). Each subpass contains the list of attachments that are written to when drawing. The RenderPass only defines the formats and dimensions of all the attachments of the multiple subpasses.

A Framebuffer defines the actual images that are attached. The format of the images must match what the RenderPass expects.

Creating a RenderPass is necessary before you create a graphics pipeline. A Framebuffer, however, is only needed when you build the command buffer.

Creating a RenderPass

Creating a RenderPass in the vulkano library is best done with the single_pass_renderpass! macro.

This macro creates an inaccessible struct which implements the RenderPassLayout trait. This trait tells vulkano what the characteristics of the renderpass are, and is also used to determine the types of the various parameters later on.

Structs

AttachmentDescription

Framebuffer

PassDependencyDescription

PassDescription

RenderPass
Defines the layout of multiple subpasses.

Subpass
Represents a subpass within a RenderPass.

Enums

ClearValue
Describes a uniform value that will be used to fill an attachment at the start of the renderpass.

LoadOp
Describes what the implementation should do with an attachment at the start of the subpass.

StoreOp
Describes what the implementation should do with an attachment after all the subpasses have completed.

Traits

CompatibleLayout
Trait implemented on renderpass layouts to check whether they are compatible with another layout.

RenderPassLayout
Types that describes the characteristics of a renderpass.

RenderPassLayoutExt