Struct vulkano::framebuffer::EmptySinglePassRenderPass

pub struct EmptySinglePassRenderPass {
    // some fields omitted
}

Implementation of RenderPass with no attachment at all and a single pass.

When you use a EmptySinglePassRenderPass, the list of attachments and clear values must be ().

Methods

impl EmptySinglePassRenderPass

fn raw(device: &Arc<Device>) -> Result<EmptySinglePassRenderPass, RenderPassCreationError>

See the docs of new().

fn new(device: &Arc<Device>) -> Arc<EmptySinglePassRenderPass>

Builds the render pass.

Panic

Panicks if the device or host ran out of memory.