Struct vulkano::framebuffer::Framebuffer

pub struct Framebuffer<L> {
    // some fields omitted
}

Methods

impl<L> Framebuffer<L>

fn new<'a, M1, M2>(renderpass: &Arc<RenderPass<L>>, dimensions: (u32, u32, u32), attachments: L::AttachmentsList) -> Result<Arc<Framebuffer<L>>, OomError> where L: RenderPassLayoutExt<'a, M1, M2>

fn is_compatible_with<R>(&self, renderpass: &Arc<RenderPass<R>>) -> bool

Returns true if this framebuffer can be used with the specified renderpass.

fn width(&self) -> u32

Returns the width of the framebuffer in pixels.

fn height(&self) -> u32

Returns the height of the framebuffer in pixels.

fn layers(&self) -> u32

Returns the number of layers (or depth) of the framebuffer.

fn renderpass(&self) -> &Arc<RenderPass<L>>

Returns the renderpass that was used to create this framebuffer.

fn attachments(&self) -> &[Arc<ImageResource>]

Returns all the resources attached to that framebuffer.