Struct vulkano::framebuffer::LayoutPassDependencyDescription

pub struct LayoutPassDependencyDescription {
    pub source_subpass: usize,
    pub destination_subpass: usize,
    pub src_stages: PipelineStages,
    pub dst_stages: PipelineStages,
    pub src_access: AccessFlagBits,
    pub dst_access: AccessFlagBits,
    pub by_region: bool,
}

Describes a dependency between two passes of a render pass.

The implementation is allowed to change the order of the passes within a render pass, unless you specify that there exists a dependency between two passes (ie. the result of one will be used as the input of another one).

Fields
source_subpass: usize

Index of the subpass that writes the data that destination_subpass is going to use.

destination_subpass: usize

Index of the subpass that reads the data that source_subpass wrote.

src_stages: PipelineStages

The pipeline stages that must be finished on the previous subpass before the destination subpass can start.

dst_stages: PipelineStages

The pipeline stages of the destination subpass that must wait for the source to be finished. Stages that are earlier of the stages specified here can start before the source is finished.

src_access: AccessFlagBits

The way the source subpass accesses the attachments on which we depend.

dst_access: AccessFlagBits

The way the destination subpass accesses the attachments on which we depend.

by_region: bool

If false, then the whole subpass must be finished for the next one to start. If true, then the implementation can start the new subpass for some given pixels as long as the previous subpass is finished for these given pixels.

In other words, if the previous subpass has some side effects on other parts of an attachment, then you sould set it to false.

Passing false is always safer than passing true, but in practice you rarely need to pass false.