Struct vulkano::framebuffer::RenderPass

pub struct RenderPass<L> {
    // some fields omitted
}

Defines the layout of multiple subpasses.

Methods

impl<L> RenderPass<L> where L: RenderPassLayout

fn new(device: &Arc<Device>, layout: L) -> Result<Arc<RenderPass<L>>, OomError>

Builds a new renderpass.

This function calls the methods of the RenderPassLayout implementation and builds the corresponding Vulkan object.

Panic

Panicks if this functions detects that the RenderPassLayout trait was not implemented correctly and contains an error. See the documentation of the various methods and structs related to RenderPassLayout for more details.

impl RenderPass<EmptySinglePassLayout>

fn empty_single_pass(device: &Arc<Device>) -> Result<Arc<RenderPass<EmptySinglePassLayout>>, OomError>

Builds a RenderPass with no attachment and a single pass.

impl<L> RenderPass<L> where L: RenderPassLayout

fn device(&self) -> &Arc<Device>

Returns the device that was used to create this renderpass.

fn num_subpasses(&self) -> u32

Returns the number of subpasses.

fn subpass(&self, id: u32) -> Option<Subpass<L>>

Returns a handle that represents a subpass of this renderpass.

fn is_compatible_with<R2>(&self, other: &RenderPass<R2>) -> bool where R2: RenderPassLayout

Returns true if this renderpass is compatible with another one.

This means that framebuffers created with this renderpass can also be used alongside with the other renderpass.

fn layout(&self) -> &L

Returns the layout used to create this renderpass.