Struct vulkano::framebuffer::Subpass

pub struct Subpass<'a, L: 'a> {
    // some fields omitted
}

Represents a subpass within a RenderPass.

This struct doesn't correspond to anything in Vulkan. It is simply an equivalent to a combination of a render pass and subpass ID.

Methods

impl<'a, L: 'a> Subpass<'a, L>

fn renderpass(&self) -> &'a RenderPass<L>

Returns the renderpass of this subpass.

fn index(&self) -> u32

Returns the index of this subpass within the renderpass.

fn has_color_or_depth_stencil_attachment(&self) -> bool

Returns true if the subpass has any color or depth/stencil attachment.

fn num_samples(&self) -> Option<u32>

Returns the number of samples in the color and/or depth/stencil attachments. Returns None if there is no such attachment in this subpass.