Struct vulkano::framebuffer::UnsafeRenderPass

pub struct UnsafeRenderPass {
    // some fields omitted
}

Defines the layout of multiple subpasses.

Methods

impl UnsafeRenderPass

unsafe fn new<Ia, Ip, Id>(device: &Arc<Device>, attachments: Ia, passes: Ip, pass_dependencies: Id) -> Result<UnsafeRenderPass, RenderPassCreationError> where Ia: ExactSizeIterator<Item=LayoutAttachmentDescription> + Clone, Ip: ExactSizeIterator<Item=LayoutPassDescription> + Clone, Id: ExactSizeIterator<Item=LayoutPassDependencyDescription>

Builds a new renderpass.

Safety

This function doesn't check whether all the restrictions in the attachments, passes and passes dependencies were enforced.

See the documentation of the structs of this module for more info about these restrictions.

Panic

Can panick if it detects some violations in the restrictions. Only unexpensive checks are performed. debug_assert! is used, so some restrictions are only checked in debug mode.

fn device(&self) -> &Arc<Device>

Returns the device that was used to create this render pass.