Trait vulkano::framebuffer::FramebufferAbstract

pub unsafe trait FramebufferAbstract: RenderPassAbstract {
    fn inner(&self) -> FramebufferSys;

    fn dimensions(&self) -> [u32; 3];

    fn attached_image_view(&self, index: usize) -> Option<&dyn ImageViewAccess>;

    fn width(&self) -> u32 { ... }

    fn height(&self) -> u32 { ... }

    fn layers(&self) -> u32 { ... }
}

Trait for objects that contain a Vulkan framebuffer object.

Any Framebuffer object implements this trait. You can therefore turn a Arc<Framebuffer<_>> into a Arc<FramebufferAbstract + Send + Sync> for easier storage.

Required methods

fn inner(&self) -> FramebufferSys

Returns an opaque struct that represents the framebuffer's internals.

fn dimensions(&self) -> [u32; 3]

Returns the width, height and array layers of the framebuffer.

fn attached_image_view(&self, index: usize) -> Option<&dyn ImageViewAccess>

Returns the attachment of the framebuffer with the given index.

If the index is not between 0 and num_attachments, then None should be returned.

Loading content...
Provided methods

fn width(&self) -> u32

Returns the width of the framebuffer in pixels.

fn height(&self) -> u32

Returns the height of the framebuffer in pixels.

fn layers(&self) -> u32

Returns the number of layers (or depth) of the framebuffer.

Loading content...