Trait vulkano::framebuffer::RenderPassAttachmentsList

pub unsafe trait RenderPassAttachmentsList<A>: RenderPass {
    type AttachmentsIter: ExactSizeIterator<Item=(Arc<ImageView>, Arc<Image>, ImageLayout, ImageLayout)>;
    fn convert_attachments_list(&self, A) -> Result<Self::AttachmentsIter, FramebufferCreationError>;
}

Extension trait for RenderPass. Defines which types are allowed as an attachments list.

Safety

This trait is unsafe because it's the job of the implementation to check whether the attachments list is correct. What needs to be checked:

That the attachments' format and samples count match the render pass layout.

That the attachments have been created with the proper usage flags.

That the attachments only expose one mipmap.

That the attachments use identity components swizzling. TODO: more stuff with aliasing

Associated Types

type AttachmentsIter: ExactSizeIterator<Item=(Arc<ImageView>, Arc<Image>, ImageLayout, ImageLayout)>

A decoded A.

Required Methods

fn convert_attachments_list(&self, A) -> Result<Self::AttachmentsIter, FramebufferCreationError>

Decodes a A into a list of attachments.

Returns an error if one of the attachments is wrong.