Trait vulkano::framebuffer::RenderPassClearValues

pub unsafe trait RenderPassClearValues<C>: RenderPass {
    type ClearValuesIter: Iterator<Item=ClearValue>;
    fn convert_clear_values(&self, C) -> Self::ClearValuesIter;
}

Extension trait for RenderPass. Defines which types are allowed as a list of clear values.

Safety

This trait is unsafe because vulkano doesn't check whether the clear value is in a format that matches the attachment.

Associated Types

type ClearValuesIter: Iterator<Item=ClearValue>

Iterator that produces one clear value per attachment.

Required Methods

fn convert_clear_values(&self, C) -> Self::ClearValuesIter

Decodes a C into a list of clear values where each element corresponds to an attachment. The size of the returned iterator must be the same as the number of attachments.

The format of the clear value must match the format of the attachment. Attachments that are not loaded with LoadOp::Clear must have an entry equal to ClearValue::None.