Trait vulkano::framebuffer::RenderPassLayout

pub unsafe trait RenderPassLayout {
    type ClearValues;
    type ClearValuesIter: Iterator<Item=ClearValue>;
    type AttachmentsDescIter: ExactSizeIterator<Item=AttachmentDescription>;
    type PassesIter: ExactSizeIterator<Item=PassDescription>;
    type PassDependenciesIter: ExactSizeIterator<Item=PassDependencyDescription>;
    type AttachmentsList;
    type AttachmentsIter: ExactSizeIterator<Item=Arc<AbstractImageView>>;
    fn convert_clear_values(&self, Self::ClearValues) -> Self::ClearValuesIter;
    fn attachments(&self) -> Self::AttachmentsDescIter;
    fn passes(&self) -> Self::PassesIter;
    fn pass_dependencies(&self) -> Self::PassDependenciesIter;
    fn convert_attachments_list(&self, Self::AttachmentsList) -> Self::AttachmentsIter;
}

Types that describes the characteristics of a renderpass.

Associated Types

type ClearValues

The list of clear values to use when beginning to draw on this renderpass.

type ClearValuesIter: Iterator<Item=ClearValue>

Iterator that produces one clear value per attachment.

type AttachmentsDescIter: ExactSizeIterator<Item=AttachmentDescription>

Iterator that produces attachments.

type PassesIter: ExactSizeIterator<Item=PassDescription>

Iterator that produces passes.

type PassDependenciesIter: ExactSizeIterator<Item=PassDependencyDescription>

Iterator that produces pass dependencies.

type AttachmentsList

List of images that will be binded to attachments.

A parameter of this type must be passed when creating a Framebuffer.

type AttachmentsIter: ExactSizeIterator<Item=Arc<AbstractImageView>>

A decoded AttachmentsList.

Required Methods

fn convert_clear_values(&self, Self::ClearValues) -> Self::ClearValuesIter

Decodes a ClearValues into a list of clear values where each element corresponds to an attachment. The size of the returned array must be the same as the number of attachments.

The format of the clear value must match the format of the attachment. Only attachments that are loaded with LoadOp::Clear must have an entry in the array.

fn attachments(&self) -> Self::AttachmentsDescIter

Returns the descriptions of the attachments.

fn passes(&self) -> Self::PassesIter

Returns the descriptions of the passes.

fn pass_dependencies(&self) -> Self::PassDependenciesIter

Returns the descriptions of the dependencies between passes.

fn convert_attachments_list(&self, Self::AttachmentsList) -> Self::AttachmentsIter

Decodes a AttachmentsList into a list of attachments.