Trait vulkano::framebuffer::RenderPassSubpassInterface

pub unsafe trait RenderPassSubpassInterface<Other: ?Sized>: RenderPassDesc where
    Other: ShaderInterfaceDef, {
    fn is_compatible_with(&self, subpass: u32, other: &Other) -> bool;
}

Extension trait for RenderPassDesc that checks whether a subpass of this render pass accepts the output of a fragment shader.

The trait is automatically implemented for all type that implement RenderPassDesc and RenderPassDesc.

Note: This trait exists so that you can specialize it once specialization lands in Rust.

Required methods

fn is_compatible_with(&self, subpass: u32, other: &Other) -> bool

Returns true if this subpass is compatible with the fragment output definition. Also returns false if the subpass is out of range.

Loading content...