Module vulkano::image

Images storage (1D, 2D, 3D, arrays, etc.).

Strong typing

Images in vulkano are strong-typed. Their signature is Image<Ty, F, M>.

The Ty parameter describes the type of image: 1D, 2D, 3D, 1D array, 2D array. All these come in two variants: with or without multisampling. The actual type of Ty must be one of the marker structs of this module that start with the Ty prefix.

The F parameter describes the format of each pixel of the image. It must be one of the marker structs of the formats module.

The M parameter describes where the image's memory was allocated from. It is similar to buffers.

Structs

Image
A storage for pixels or arbitrary data.

ImagePrototype
Prototype of an image.

ImageView
A representation of an image.

Swizzle

Type1d

Type1dArray

Type1dArrayMultisample

Type1dMultisample

Type2d

Type2dArray

Type2dArrayMultisample

Type2dMultisample

Type3d

Type3dMultisample

TypeCube

TypeCubeArray

TypeCubeArrayMultisample

TypeCubeMultisample

Usage
Describes how an image is going to be used. This is not an optimization.

Enums

ComponentSwizzle

ImageType

Layout

MipmapsCount
Specifies how many mipmaps must be allocated.

Traits

AbstractImage

AbstractImageView

AbstractTypedImageView

AbstractTypedImageViewsTuple

CanCreateView

ImageTypeMarker

ImageViewTypeMarker

MultisampleType

TypeMarker