Struct vulkano::image::attachment::AttachmentImage

pub struct AttachmentImage<F, A = StdMemoryPool> where A: MemoryPool {
    // some fields omitted
}

Image whose purpose is to be used as a framebuffer attachment.

Methods

impl<F> AttachmentImage<F>

fn new(device: &Arc<Device>, dimensions: [u32; 2], format: F) -> Result<Arc<AttachmentImage<F>>, ImageCreationError> where F: FormatDesc

Creates a new image with the given dimensions and format.

Returns an error if the dimensions are too large or if the backend doesn't support this format as a framebuffer attachment.

fn transient(device: &Arc<Device>, dimensions: [u32; 2], format: F) -> Result<Arc<AttachmentImage<F>>, ImageCreationError> where F: FormatDesc

Same as new, except that the image will be transient.

A transient image is special because its content is undefined outside of a render pass. This means that the implementation has the possibility to not allocate any memory for it.

impl<F, A> AttachmentImage<F, A> where A: MemoryPool

fn dimensions(&self) -> [u32; 2]

Returns the dimensions of the image.