Struct vulkano::image::immutable::ImmutableImage

pub struct ImmutableImage<F, A = StdMemoryPool> where A: MemoryPool {
    // some fields omitted
}

Image whose purpose is to be used for read-only purposes. You can write to the image once, but then you must only ever read from it. TODO: clarify because of blit operations

Methods

impl<F> ImmutableImage<F>

fn new<'a, I>(device: &Arc<Device>, dimensions: Dimensions, format: F, queue_families: I) -> Result<Arc<ImmutableImage<F>>, ImageCreationError> where F: FormatDesc, I: IntoIterator<Item=QueueFamily<'a>>

impl<F, A> ImmutableImage<F, A> where A: MemoryPool

fn dimensions(&self) -> Dimensions