Struct vulkano::image::ImagePrototype

pub struct ImagePrototype<Ty, F, M> where Ty: ImageTypeMarker {
    // some fields omitted
}

Prototype of an image.

Needs to be transitionned to a proper layout in order to be turned into a regular Image.

Methods

impl<Ty, F, M> ImagePrototype<Ty, F, M> where M: MemorySourceChunk, Ty: ImageTypeMarker, F: FormatDesc

fn dimensions(&self) -> Ty::Dimensions

Returns the dimensions of this image.

fn array_layers(&self) -> u32

Returns the number of array layers of this image.

fn mipmap_levels(&self) -> u32

Returns the number of mipmap levels of this image.

fn num_samples(&self) -> u32

Returns the number of samples of each pixel of this image.

Returns 1 if the image is not multisampled.

fn transition(self, layout: Layout, pool: &CommandBufferPool, submit_queue: &Arc<Queue>) -> Result<Arc<Image<Ty, F, M>>, OomError>

Transitions the image prototype into a real image by submitting a one-shot command buffer.

Panic

Panicks if layout is Undefined or Preinitialized.