Struct vulkano::image::ImageView

pub struct ImageView<Ty, F, M> where Ty: ImageTypeMarker {
    // some fields omitted
}

A representation of an image.

Accessing an image from within a shader can only be done through an ImageView. An ImageView represents a region of an image. You can also do things like creating a 2D view of a 3D image, swizzle the channels, or change the format of the texture (with some restrictions).

Methods

impl<Ty, F, M> ImageView<Ty, F, M> where Ty: ImageTypeMarker

fn new(image: &Arc<Image<Ty, F, M>>) -> Result<Arc<ImageView<Ty, F, M>>, OomError> where F: FormatDesc

Creates a new view from an image.

Note that you must create the view with identity swizzling if you want to use this view as a framebuffer attachment.

fn image(&self) -> &Arc<Image<Ty, F, M>>

Returns the image from which this view is taken from.

fn is_identity_swizzled(&self) -> bool

Returns true if the swizzling of this image view is identity.