Struct vulkano::image::StorageImage

pub struct StorageImage<F, A = Arc<StdMemoryPool>> where
    A: MemoryPool,  { /* fields omitted */ }

General-purpose image in device memory. Can be used for any usage, but will be slower than a specialized image.

Implementations

impl<F> StorageImage<F>

pub fn new<'a, I>(
device: Arc<Device>,
dimensions: Dimensions,
format: F,
queue_families: I
) -> Result<Arc<StorageImage<F>>, ImageCreationError> where
F: FormatDesc,
I: IntoIterator<Item = QueueFamily<'a>>,

Creates a new image with the given dimensions and format.

pub fn with_usage<'a, I>(
device: Arc<Device>,
dimensions: Dimensions,
format: F,
usage: ImageUsage,
queue_families: I
) -> Result<Arc<StorageImage<F>>, ImageCreationError> where
F: FormatDesc,
I: IntoIterator<Item = QueueFamily<'a>>,

Same as new, but allows specifying the usage.

impl<F, A> StorageImage<F, A> where
A: MemoryPool,

pub fn dimensions(&self) -> Dimensions

Returns the dimensions of the image.