Struct vulkano::image::swapchain::SwapchainImage

pub struct SwapchainImage<W> { /* fields omitted */ }

An image that is part of a swapchain.

Creating a SwapchainImage is automatically done when creating a swapchain.

A swapchain image is special in the sense that it can only be used after being acquired by calling the acquire method on the swapchain. You have no way to know in advance which swapchain image is going to be acquired, so you should keep all of them alive.

After a swapchain image has been acquired, you are free to perform all the usual operations on it. When you are done you can then present the image (by calling the corresponding method on the swapchain), which will have the effect of showing the content of the image to the screen. Once an image has been presented, it can no longer be used unless it is acquired again.

Methods

impl<W> SwapchainImage<W>

pub unsafe fn from_raw(
swapchain: Arc<Swapchain<W>>,
id: usize
) -> Result<Arc<SwapchainImage<W>>, OomError>

Builds a SwapchainImage from raw components.

This is an internal method that you shouldn't call.

pub fn dimensions(&self) -> [u32; 2]

Returns the dimensions of the image.

A SwapchainImage is always two-dimensional.

pub fn swapchain(&self) -> &Arc<Swapchain<W>>

Returns the swapchain this image belongs to.