Struct vulkano::image::sys::UnsafeImageView

pub struct UnsafeImageView { /* fields omitted */ }

Implementations

impl UnsafeImageView

pub unsafe fn raw(
image: &UnsafeImage,
ty: ViewType,
mipmap_levels: Range<u32>,
array_layers: Range<u32>
) -> Result<UnsafeImageView, OomError>

See the docs of new().

pub unsafe fn new(
image: &UnsafeImage,
ty: ViewType,
mipmap_levels: Range<u32>,
array_layers: Range<u32>
) -> UnsafeImageView

Creates a new view from an image.

Note that you must create the view with identity swizzling if you want to use this view as a framebuffer attachment.

Panic

Panics if mipmap_levels or array_layers is out of range of the image.

Panics if the view types doesn't match the dimensions of the image (for example a 2D view from a 3D image).

Panics if trying to create a cubemap with a number of array layers different from 6.

Panics if trying to create a cubemap array with a number of array layers not a multiple of 6.

Panics if the device or host ran out of memory.

pub fn format(&self) -> Format

pub fn usage_transfer_source(&self) -> bool

pub fn usage_transfer_destination(&self) -> bool

pub fn usage_sampled(&self) -> bool

pub fn usage_storage(&self) -> bool

pub fn usage_color_attachment(&self) -> bool

pub fn usage_depth_stencil_attachment(&self) -> bool

pub fn usage_transient_attachment(&self) -> bool

pub fn usage_input_attachment(&self) -> bool