Trait vulkano::image::AbstractImageView

pub unsafe trait AbstractImageView: Resource + VulkanObjectU64 {
    fn default_layout(&self) -> Layout;
    unsafe fn gpu_access(&self, write: bool, queue: &Arc<Queue>, fence: Option<Arc<Fence>>, semaphore: Option<Arc<Semaphore>>) -> Option<Arc<Semaphore>>;
    fn usage_transfer_src(&self) -> bool;
    fn usage_transfer_dest(&self) -> bool;
    fn usage_sampled(&self) -> bool;
    fn usage_storage(&self) -> bool;
    fn usage_color_attachment(&self) -> bool;
    fn usage_depth_stencil_attachment(&self) -> bool;
    fn usage_transient_attachment(&self) -> bool;
    fn usage_input_attachment(&self) -> bool;
}

Required Methods

fn default_layout(&self) -> Layout

unsafe fn gpu_access(&self, write: bool, queue: &Arc<Queue>, fence: Option<Arc<Fence>>, semaphore: Option<Arc<Semaphore>>) -> Option<Arc<Semaphore>>

fn usage_transfer_src(&self) -> bool

True if the image can be used as a source for transfers.

fn usage_transfer_dest(&self) -> bool

True if the image can be used as a destination for transfers.

fn usage_sampled(&self) -> bool

True if the image can be sampled from a shader.

fn usage_storage(&self) -> bool

True if the image can be used for image loads/stores in shaders.

fn usage_color_attachment(&self) -> bool

True if the image can be used as a color attachment in a framebuffer.

fn usage_depth_stencil_attachment(&self) -> bool

True if the image can be used as a depth and/or stencil attachment in a framebuffer.

fn usage_transient_attachment(&self) -> bool

True if the image can be used as a transient attachment in a framebuffer.

fn usage_input_attachment(&self) -> bool

True if the image can be used as an input attachment in a framebuffer.