Module vulkano::image::traits

Structs

ImageAccessFromUndefinedLayout
Wraps around an object that implements ImageAccess and modifies the initial layout requirement to be either Undefined or Preinitialized.

ImageInner
Inner information about an image.

Traits

AttachmentImageView

ImageAccess
Trait for types that represent the way a GPU can access an image.

ImageClearValue
Extension trait for images. Checks whether the value T can be used as a clear value for the given image.

ImageContent

ImageViewAccess
Trait for types that represent the GPU can access an image view.