Trait vulkano::image::traits::ImageView

pub unsafe trait ImageView: 'static + Send + Sync {
    fn parent(&self) -> &Image;
    fn parent_arc(&Arc<Self>) -> Arc<Image> where Self: Sized;
    fn inner(&self) -> &UnsafeImageView;
    fn blocks(&self) -> Vec<(u32, u32)>;
    fn descriptor_set_storage_image_layout(&self) -> Layout;
    fn descriptor_set_combined_image_sampler_layout(&self) -> Layout;
    fn descriptor_set_sampled_image_layout(&self) -> Layout;
    fn descriptor_set_input_attachment_layout(&self) -> Layout;
    fn identity_swizzle(&self) -> bool;

    fn format(&self) -> Format { ... }
    fn samples(&self) -> u32 { ... }
    fn can_be_sampled(&self, sampler: &Sampler) -> bool { ... }
}

Trait for types that represent image views.

Required Methods

fn parent(&self) -> &Image

fn parent_arc(&Arc<Self>) -> Arc<Image> where Self: Sized

fn inner(&self) -> &UnsafeImageView

Returns the inner unsafe image view object used by this image view.

fn blocks(&self) -> Vec<(u32, u32)>

Returns the blocks of the parent image this image view overlaps.

fn descriptor_set_storage_image_layout(&self) -> Layout

Returns the image layout to use in a descriptor with the given subresource.

fn descriptor_set_combined_image_sampler_layout(&self) -> Layout

Returns the image layout to use in a descriptor with the given subresource.

fn descriptor_set_sampled_image_layout(&self) -> Layout

Returns the image layout to use in a descriptor with the given subresource.

fn descriptor_set_input_attachment_layout(&self) -> Layout

Returns the image layout to use in a descriptor with the given subresource.

fn identity_swizzle(&self) -> bool

Returns true if the view doesn't use components swizzling.

Must be true when the view is used as a framebuffer attachment or TODO: I don't remember the other thing.

Provided Methods

fn format(&self) -> Format

Returns the format of this view. This can be different from the parent's format.

fn samples(&self) -> u32

fn can_be_sampled(&self, sampler: &Sampler) -> bool

Returns true if the given sampler can be used with this image view.

This method should check whether the sampler's configuration can be used with the format of the view.