Trait vulkano::image::traits::ImageViewAccess

pub unsafe trait ImageViewAccess {
    fn parent(&self) -> &dyn ImageAccess;

    fn dimensions(&self) -> Dimensions;

    fn inner(&self) -> &UnsafeImageView;

    fn descriptor_set_storage_image_layout(&self) -> ImageLayout;

    fn descriptor_set_combined_image_sampler_layout(&self) -> ImageLayout;

    fn descriptor_set_sampled_image_layout(&self) -> ImageLayout;

    fn descriptor_set_input_attachment_layout(&self) -> ImageLayout;

    fn identity_swizzle(&self) -> bool;

    fn format(&self) -> Format { ... }

    fn samples(&self) -> u32 { ... }

    fn can_be_sampled(&self, _sampler: &Sampler) -> bool { ... }
}

Trait for types that represent the GPU can access an image view.

Required methods

fn parent(&self) -> &dyn ImageAccess

fn dimensions(&self) -> Dimensions

Returns the dimensions of the image view.

fn inner(&self) -> &UnsafeImageView

Returns the inner unsafe image view object used by this image view.

fn descriptor_set_storage_image_layout(&self) -> ImageLayout

Returns the image layout to use in a descriptor with the given subresource.

fn descriptor_set_combined_image_sampler_layout(&self) -> ImageLayout

Returns the image layout to use in a descriptor with the given subresource.

fn descriptor_set_sampled_image_layout(&self) -> ImageLayout

Returns the image layout to use in a descriptor with the given subresource.

fn descriptor_set_input_attachment_layout(&self) -> ImageLayout

Returns the image layout to use in a descriptor with the given subresource.

fn identity_swizzle(&self) -> bool

Returns true if the view doesn't use components swizzling.

Must be true when the view is used as a framebuffer attachment or TODO: I don't remember the other thing.

Loading content...
Provided methods

fn format(&self) -> Format

Returns the format of this view. This can be different from the parent's format.

fn samples(&self) -> u32

fn can_be_sampled(&self, _sampler: &Sampler) -> bool

Returns true if the given sampler can be used with this image view.

This method should check whether the sampler's configuration can be used with the format of the view.

Loading content...