Struct vulkano::instance::debug::MessageTypes

pub struct MessageTypes {
    pub error: bool,
    pub warning: bool,
    pub performance_warning: bool,
    pub information: bool,
    pub debug: bool,
}

Type of message.

Fields
error: bool

An error that may cause undefined results, including an application crash.

warning: bool

An unexpected use.

performance_warning: bool

A potential non-optimal use.

information: bool

An informational message that may be handy when debugging an application.

debug: bool

Diagnostic information from the loader and layers.

Methods

impl MessageTypes

fn errors() -> MessageTypes

Builds a MessageTypes with all fields set to false expect error.

fn errors_and_warnings() -> MessageTypes

Builds a MessageTypes with all fields set to false expect error, warning and performance_warning.

fn none() -> MessageTypes

Builds a MessageTypes with all fields set to false.