Struct vulkano::instance::ApplicationInfo

pub struct ApplicationInfo<'a> {
    pub application_name: &'a str,
    pub application_version: u32,
    pub engine_name: &'a str,
    pub engine_version: u32,
}

Information that can be given to the Vulkan driver so that it can identify your application.

Fields

application_name

Name of the application.

application_version

An opaque number that contains the version number of the application.

engine_name

Name of the engine used to power the application.

engine_version

An opaque number that contains the version number of the engine.