Struct vulkano::instance::MemoryType

pub struct MemoryType<'a> {
    // some fields omitted
}

Represents a memory type in a physical device.

Methods

impl<'a> MemoryType<'a>

fn physical_device(&self) -> PhysicalDevice<'a>

Returns the physical device associated to this memory type.

fn id(&self) -> u32

Returns the identifier of this memory type within the physical device.

fn heap(&self) -> MemoryHeap<'a>

Returns the heap that corresponds to this memory type.

fn is_device_local(&self) -> bool

Returns true if the memory type is located on the device, which means that it's the most efficient for GPU accesses.

fn is_host_visible(&self) -> bool

Returns true if the memory type can be accessed by the host.

fn is_host_coherent(&self) -> bool

Returns true if modifications made by the host or the GPU on this memory type are instantaneously visible to the other party. False means that changes have to be flushed.

You don't need to worry about this, as this library handles that for you.

fn is_host_cached(&self) -> bool

Returns true if memory of this memory type is cached by the host. Host memory accesses to cached memory is faster than for uncached memory. However you are not guaranteed that it is coherent.

fn is_lazily_allocated(&self) -> bool

Returns true if allocations made to this memory type is lazy.

This means that no actual allocation is performed. Instead memory is automatically allocated by the Vulkan implementation.

Memory of this type can only be used on images created with a certain flag. Memory of this type is never host-visible.