Struct vulkano::instance::QueueFamily

pub struct QueueFamily<'a> {
    // some fields omitted
}

Represents a queue family in a physical device.

A queue family is group of one or multiple queues. All queues of one family have the same characteristics.

Methods

impl<'a> QueueFamily<'a>

fn physical_device(&self) -> PhysicalDevice<'a>

Returns the physical device associated to this queue family.

fn id(&self) -> u32

Returns the identifier of this queue family within the physical device.

fn queues_count(&self) -> usize

Returns the number of queues that belong to this family.

Guaranteed to be at least 1 (or else that family wouldn't exist).

fn supports_graphics(&self) -> bool

Returns true if queues of this family can execute graphics operations.

fn supports_compute(&self) -> bool

Returns true if queues of this family can execute compute operations.

fn supports_transfers(&self) -> bool

Returns true if queues of this family can execute transfer operations.

fn supports_sparse_binding(&self) -> bool

Returns true if queues of this family can execute sparse resources binding operations.