Module vulkano::memory

GPU-visible memory allocation and management.

When you create a buffer or a texture with Vulkan, you have to bind it to a chunk of allocated memory. To do so, you have to pass a type that implements the MemorySource trait.

There are several implementations of the trait, ie. several things that you can pass to the constructors of buffers and textures:

&Arc<Device>, which will simply allocate a new chunk of memory every time (easy but not very efficient).

MemorySource, which is the same as &Arc<Device> except that it will use the already-allocated block.

... needs more ...

Synchronization

In Vulkan, it's the job of the programmer to enforce memory safety. In other words, the programmer must take care that two chunks of memory are not read and written simultaneously.

In this library, this is enforced by the implementation of MemorySource or MemorySourceChunk.

There are two mechanisms in Vulkan that can provide synchronization: fences and semaphores. Fences provide synchronization between the CPU and the GPU, and semaphores provide synchronization between multiple queues of the GPU. See the sync module for more info.

Sparse resources

Not yet implemented.

Instead of creating a buffer or an image with a single chunk of memory, you also have the possibility to create resources with sparse memory.

For example you can bind the first half of the buffer to a memory chunk, and the second half of the buffer to another memory chunk.

There is a hierarchy of three features related to sparse resources:

The sparseBinding feature allows you to use sparse resources.

The sparseResidency feature is a superset of sparseBinding and allows you to leave some parts of the resource unbinded before using it.

The sparseResidencyAliased feature is a superset of sparseResidency and allows you to bind the same memory chunk to multiple different resources at once.

Structs

DeviceLocal
Dummy marker whose strategy is to allocate a new chunk of memory for each allocation.

DeviceLocalChunk
A chunk allocated from a DeviceLocal.

DeviceMemory
Represents memory that has been allocated.

HostVisible
Dummy marker whose strategy is to allocate a new chunk of memory for each allocation.

HostVisibleChunk
A chunk allocated from a HostVisible.

MappedDeviceMemory
Represents memory that has been allocated and mapped in CPU accessible space.

Enums

ChunkProperties

ChunkRange

Traits

Content
Trait for types of data that can be mapped.

CpuAccessible
Trait for memory objects that can be accessed from the CPU.

CpuWriteAccessible
Trait for memory objects that be mutably accessed from the CPU.

MemorySource
Trait for objects that can be used to fill the memory requirements of a buffer or an image.

MemorySourceChunk
A chunk of GPU-visible memory.