Trait vulkano::memory::pool::MemoryPoolAlloc

pub unsafe trait MemoryPoolAlloc {
    fn mapped_memory(&self) -> Option<&MappedDeviceMemory>;

    fn memory(&self) -> &DeviceMemory;

    fn offset(&self) -> usize;
}

Object that represents a single allocation. Its destructor should free the chunk.

Required methods

fn mapped_memory(&self) -> Option<&MappedDeviceMemory>

Returns the memory object from which this is allocated. Returns None if the memory is not mapped.

fn memory(&self) -> &DeviceMemory

Returns the memory object from which this is allocated.

fn offset(&self) -> usize

Returns the offset at the start of the memory where the first byte of this allocation resides.

Loading content...