Struct vulkano::memory::DeviceMemory

pub struct DeviceMemory {
    // some fields omitted
}

Represents memory that has been allocated.

Methods

impl DeviceMemory

fn alloc(device: &Arc<Device>, memory_type: &MemoryType, size: usize) -> Result<DeviceMemory, OomError>

Allocates a chunk of memory from the device.

Panic

Panicks if memory_type doesn't belong to the same physical device as device.

fn alloc_and_map(device: &Arc<Device>, memory_type: &MemoryType, size: usize) -> Result<MappedDeviceMemory, OomError>

Allocates a chunk of memory and maps it.

Panic

Panicks if memory_type doesn't belong to the same physical device as device.

Panicks if the memory type is not host-visible.

fn memory_type(&self) -> MemoryType

Returns the memory type this chunk was allocated on.

fn size(&self) -> usize

Returns the size in bytes of that memory chunk.

fn device(&self) -> &Arc<Device>

Returns the device associated with this allocation.