Struct vulkano::memory::MappedDeviceMemory

pub struct MappedDeviceMemory {
    // some fields omitted
}

Represents memory that has been allocated and mapped in CPU accessible space.

Methods

impl MappedDeviceMemory

fn memory(&self) -> &DeviceMemory

Returns the underlying DeviceMemory.

fn mapping_pointer(&self) -> *mut c_void

Returns a pointer to the mapping.

Note that access to this pointer is not safe at all.