Trait vulkano::memory::Content

pub unsafe trait Content {
    fn ref_from_ptr<'a>(ptr: *mut c_void, size: usize) -> Option<*mut Self>;

    fn is_size_suitable(_: usize) -> bool;

    fn indiv_size() -> usize;
}

Trait for types of data that can be mapped.

Required methods

fn ref_from_ptr<'a>(ptr: *mut c_void, size: usize) -> Option<*mut Self>

Builds a pointer to this type from a raw pointer.

fn is_size_suitable(_: usize) -> bool

Returns true if the size is suitable to store a type like this.

fn indiv_size() -> usize

Returns the size of an individual element.

Loading content...