Trait vulkano::memory::MemorySource

pub unsafe trait MemorySource {
    type Chunk: MemorySourceChunk;
    fn is_sparse(&self) -> bool;
    fn allocate(self, &Arc<Device>, size: usize, alignment: usize, memory_type_bits: u32) -> Result<Self::Chunk, OomError>;
}

Trait for objects that can be used to fill the memory requirements of a buffer or an image.

Associated Types

type Chunk: MemorySourceChunk

An object that represents one block of allocation. Returned by allocate.

Required Methods

fn is_sparse(&self) -> bool

Returns true if the chunks allocated by this source will use sparse memory.

fn allocate(self, &Arc<Device>, size: usize, alignment: usize, memory_type_bits: u32) -> Result<Self::Chunk, OomError>

Allocates a block of memory to be used.

memory_type_bits is a bitsfield which indicates from which memory type the memory can be allocated. For example if the bit 2 is set (memory_type_bits & (1 << 2) != 0), that means that the memory type whose ID is 2 can be used.

The implementation is allowed to return a chunk with a larger size or alignment.