Trait vulkano::memory::MemorySourceChunk

pub unsafe trait MemorySourceChunk {
    fn properties(&self) -> ChunkProperties;
    unsafe fn gpu_access(&self, write: bool, range: ChunkRange, queue: &Arc<Queue>, fence: Option<Arc<Fence>>, semaphore: Option<Arc<Semaphore>>) -> Option<Arc<Semaphore>>;
    fn may_alias(&self) -> bool;

    fn requires_fence(&self) -> bool { ... }
    fn requires_semaphore(&self) -> bool { ... }
}

A chunk of GPU-visible memory.

Required Methods

fn properties(&self) -> ChunkProperties

Returns the properties of this chunk.

unsafe fn gpu_access(&self, write: bool, range: ChunkRange, queue: &Arc<Queue>, fence: Option<Arc<Fence>>, semaphore: Option<Arc<Semaphore>>) -> Option<Arc<Semaphore>>

Instructs the manager that a part of this chunk of memory is going to be used by the GPU soon in the future. The function should block if the memory is currently being accessed by the CPU.

write indicates whether the GPU will write to the memory. If false, then it will only be written.

range indicates the part of the chunk that is concerned.

queue is the queue where the command buffer that accesses the memory will be submitted. If the gpu_access function submits something to that queue, it will thus be submitted beforehand. This behavior can be used for example to submit sparse binding commands.

fence is a fence that will be signaled when this GPU access will stop. It should be waited upon whenever the user wants to read this memory from the CPU. If requires_fence returned false, then this value will be None.

semaphore is a semaphore that will be signaled when this GPU access will stop. This value is intended to be returned later, in a follow-up call to gpu_access. If requires_semaphore returned false, then this value will be None.

The manager must track whether this chunk of memory is being accessed by the CPU/GPU and return a semaphore that must be waited upon by the GPU before the access can start. The semaphore being returned is usually one that has been previously passed to this function, but it doesn't need to be the case.

fn may_alias(&self) -> bool

Returns true if this chunk of memory may be used, now or in the future, by multiple buffers or images (or a combination of both) simultaneously. If you're not sure, it's safer to return true.

If this value is true, then the Vulkan implementation must be more conservative about reordering subpasses in a renderpass.

Provided Methods

fn requires_fence(&self) -> bool

Returns true if the gpu_access function should be passed a fence.

fn requires_semaphore(&self) -> bool

Returns true if the gpu_access function should be passed a semaphore.