Module vulkano::pipeline::blend

Defines how the color output of the fragment shader is written to the attachment.

Blending in details

There are three kinds of color attachments for the purpose of blending:

Attachments with a floating-point, fixed point format.

Attachments with a (non-normalized) integer format.

Attachments with a normalized integer format.

For floating-point and fixed-point formats, the blending operation is applied. For integer formats, the logic operation is applied. For normalized integer formats, the logic operation will take precedence if it is activated, otherwise the blending operation is applied.

Structs

AttachmentBlend
Describes how the blending system should behave for an individual attachment.

Blend
Describes how the color output of the fragment shader is written to the attachment. See the documentation of the blend module for more info.

Enums

AttachmentsBlend
Describes how the blending system should behave.

BlendFactor

BlendOp

LogicOp
Which logical operation to apply to the output values.