Module vulkano::pipeline::depth_stencil

Depth and stencil operations description.

After the fragment shader has finished running, each fragment goes through the depth and stencil tests.

The depth test passes of fails depending on how the depth value of each fragment compares to the existing depth value in the depth buffer at that fragment's location. Depth values are always between 0.0 and 1.0.

The stencil test passes or fails depending on how a reference value compares to the existing value in the stencil buffer at each fragment's location. Depending on the outcome of the depth and stencil tests, the value of the stencil buffer at that location can be updated.

Structs

DepthStencil
Configuration of the depth and stencil tests.

DynamicStencilValue
Container for dynamic StencilFaceFlags and value

Stencil
Configuration of a stencil test.

Enums

Compare
Specifies how two values should be compared to decide whether a test passes or fails.

DepthBounds
Allows you to ask the GPU to exclude fragments that are outside of a certain range.

StencilFaceFlags
Enum to specify which stencil state to use

StencilOp
Operation to perform after the depth and stencil tests.