Struct vulkano::pipeline::depth_stencil::DepthStencil

pub struct DepthStencil {
    pub depth_compare: Compare,
    pub depth_write: bool,
    pub depth_bounds_test: DepthBounds,
    pub stencil_front: Stencil,
    pub stencil_back: Stencil,
}

Configuration of the depth and stencil tests.

Fields
depth_compare: Compare

Comparison to use between the depth value of each fragment and the depth value currently in the depth buffer.

depth_write: bool

If true, then the value in the depth buffer will be updated when the depth test succeeds.

depth_bounds_test: DepthBounds

Allows you to ask the GPU to exclude fragments that are outside of a certain range. This is done in addition to the regular depth test.

stencil_front: Stencil

Stencil operations to use for points, lines and triangles whose front is facing the user.

stencil_back: Stencil

Stencil operations to use for triangles whose back is facing the user.

Methods

impl DepthStencil

fn disabled() -> DepthStencil

Creates a DepthStencil where both the depth and stencil tests are disabled and have no effect.

fn simple_depth_test() -> DepthStencil

Creates a DepthStencil with a Less depth test, depth_write set to true, and stencil testing disabled.