Struct vulkano::pipeline::raster::Rasterization

pub struct Rasterization {
    pub depth_clamp: bool,
    pub rasterizer_discard: bool,
    pub polygon_mode: PolygonMode,
    pub cull_mode: CullMode,
    pub front_face: FrontFace,
    pub line_width: Option<f32>,
    pub depth_bias: DepthBiasControl,
}

State of the rasterizer.

Fields
depth_clamp: bool

If true, then the depth value of the vertices will be clamped to [0.0 ; 1.0]. If false, fragments whose depth is outside of this range will be discarded.

rasterizer_discard: bool

If true, all the fragments will be discarded. This is usually used when your vertex shader has some side effects and you don't need to run the fragment shader.

polygon_mode: PolygonMode

This setting can ask the rasterizer to downgrade triangles into lines or points, or lines into points.

cull_mode: CullMode

Specifies whether front faces or back faces should be discarded, or none, or both.

front_face: FrontFace

Specifies which triangle orientation corresponds to the front or the triangle.

line_width: Option<f32>

Width, in pixels, of lines when drawing lines.

If you pass None, then this state will be considered as dynamic and the line width will need to be set when you build the command buffer.

depth_bias: DepthBiasControl