Trait vulkano::pipeline::shader::GraphicsEntryPointAbstract

pub unsafe trait GraphicsEntryPointAbstract: EntryPointAbstract {
    type InputDefinition: ShaderInterfaceDef;
    type OutputDefinition: ShaderInterfaceDef;
    fn input(&self) -> &Self::InputDefinition;

    fn output(&self) -> &Self::OutputDefinition;

    fn ty(&self) -> GraphicsShaderType;
}

Associated Types

type InputDefinition: ShaderInterfaceDef

type OutputDefinition: ShaderInterfaceDef

Loading content...
Required methods

fn input(&self) -> &Self::InputDefinition

Returns the input attributes used by the shader stage.

fn output(&self) -> &Self::OutputDefinition

Returns the output attributes used by the shader stage.

fn ty(&self) -> GraphicsShaderType

Returns the type of shader.

Loading content...