Trait vulkano::pipeline::shader::ShaderInterfaceDef

pub unsafe trait ShaderInterfaceDef {
    type Iter: ExactSizeIterator<Item = ShaderInterfaceDefEntry>;
    fn elements(&self) -> Self::Iter;
}

Types that contain the definition of an interface between two shader stages, or between the outside and a shader stage.

Safety

Must only provide one entry per location.

The format of each element must not be larger than 128 bits.

Associated Types

type Iter: ExactSizeIterator<Item = ShaderInterfaceDefEntry>

Iterator returned by elements.

Loading content...
Required methods

fn elements(&self) -> Self::Iter

Iterates over the elements of the interface.

Loading content...