Struct vulkano::pipeline::ComputePipeline

pub struct ComputePipeline<Pl> { /* fields omitted */ }

A pipeline object that describes to the Vulkan implementation how it should perform compute operations.

The template parameter contains the descriptor set to use with this pipeline.

All compute pipeline objects implement the ComputePipelineAbstract trait. You can turn any Arc<ComputePipeline<Pl>> into an Arc<ComputePipelineAbstract> if necessary.

Methods

impl ComputePipeline<()>

pub fn new<Cs>(
device: Arc<Device>,
shader: &Cs,
specialization: &Cs::SpecializationConstants
) -> Result<ComputePipeline<PipelineLayout<Cs::PipelineLayout>>, ComputePipelineCreationError> where
Cs::PipelineLayout: Clone,
Cs: EntryPointAbstract,

Builds a new ComputePipeline.

impl<Pl> ComputePipeline<Pl>

pub fn with_pipeline_layout<Cs>(
device: Arc<Device>,
shader: &Cs,
specialization: &Cs::SpecializationConstants,
pipeline_layout: Pl
) -> Result<ComputePipeline<Pl>, ComputePipelineCreationError> where
Cs::PipelineLayout: Clone,
Cs: EntryPointAbstract,
Pl: PipelineLayoutAbstract,

Builds a new ComputePipeline with a specific pipeline layout.

An error will be returned if the pipeline layout isn't a superset of what the shader uses.

pub unsafe fn with_unchecked_pipeline_layout<Cs>(
device: Arc<Device>,
shader: &Cs,
specialization: &Cs::SpecializationConstants,
pipeline_layout: Pl
) -> Result<ComputePipeline<Pl>, ComputePipelineCreationError> where
Cs::PipelineLayout: Clone,
Cs: EntryPointAbstract,
Pl: PipelineLayoutAbstract,

Same as with_pipeline_layout, but doesn't check whether the pipeline layout is a superset of what the shader expects.

impl<Pl> ComputePipeline<Pl>

pub fn device(&self) -> &Arc<Device>

Returns the Device this compute pipeline was created with.

pub fn layout(&self) -> &Pl

Returns the pipeline layout used in this compute pipeline.