Struct vulkano::pipeline::GraphicsPipeline

pub struct GraphicsPipeline<MultiVertex, Layout> {
    // some fields omitted
}

The template parameter contains the descriptor set to use with this pipeline, and the renderpass layout.

Methods

impl<MV, L> GraphicsPipeline<MV, L> where MV: MultiVertex, L: PipelineLayoutDesc

fn new<Vi, Fo, R, Vl, Fl>(device: &Arc<Device>, vertex_shader: &VertexShaderEntryPoint<Vi, Vl>, input_assembly: &InputAssembly, viewport: &ViewportsState, raster: &Rasterization, multisample: &Multisample, blend: &Blend, fragment_shader: &FragmentShaderEntryPoint<Fo, Fl>, layout: &Arc<PipelineLayout<L>>, render_pass: &Subpass<R>) -> Result<Arc<GraphicsPipeline<MV, L>>, OomError>

Builds a new graphics pipeline object.

Panic

Panicks if primitive restart is enabled and the topology doesn't support this feature.

Panicks if the rasterization_samples parameter of multisample is not >= 1.

Panicks if the sample_shading parameter of multisample is not between 0.0 and 1.0.

Panicks if the line width is different from 1.0 and the wide_lines feature is not enabled.

impl<MV, L> GraphicsPipeline<MV, L> where MV: MultiVertex, L: PipelineLayoutDesc

fn layout(&self) -> &Arc<PipelineLayout<L>>

Returns the pipeline layout used in the constructor.

impl<MultiVertex, Layout> GraphicsPipeline<MultiVertex, Layout>

fn has_dynamic_line_width(&self) -> bool

Returns true if the line width used by this pipeline is dynamic.