Trait vulkano::pipeline::GraphicsPipelineAbstract

pub unsafe trait GraphicsPipelineAbstract: PipelineLayoutAbstract + RenderPassAbstract + VertexSource<Vec<Arc<dyn BufferAccess + Send + Sync>>> {
    fn inner(&self) -> GraphicsPipelineSys;

    fn subpass_index(&self) -> u32;

    fn has_dynamic_line_width(&self) -> bool;

    fn num_viewports(&self) -> u32;

    fn has_dynamic_viewports(&self) -> bool;

    fn has_dynamic_scissors(&self) -> bool;

    fn has_dynamic_depth_bounds(&self) -> bool;

    fn has_dynamic_stencil_compare_mask(&self) -> bool;

    fn has_dynamic_stencil_write_mask(&self) -> bool;

    fn has_dynamic_stencil_reference(&self) -> bool;

    fn subpass(self) -> Subpass<Self>
    where
        Self: Sized,
    { ... }
}

Trait implemented on objects that reference a graphics pipeline. Can be made into a trait object. When using this trait AutoCommandBufferBuilder::draw* calls will need the buffers to be wrapped in a vec!().

Required methods

fn inner(&self) -> GraphicsPipelineSys

Returns an opaque object that represents the inside of the graphics pipeline.

fn subpass_index(&self) -> u32

Returns the index of the subpass this graphics pipeline is rendering to.

fn has_dynamic_line_width(&self) -> bool

Returns true if the line width used by this pipeline is dynamic.

fn num_viewports(&self) -> u32

Returns the number of viewports and scissors of this pipeline.

fn has_dynamic_viewports(&self) -> bool

Returns true if the viewports used by this pipeline are dynamic.

fn has_dynamic_scissors(&self) -> bool

Returns true if the scissors used by this pipeline are dynamic.

fn has_dynamic_depth_bounds(&self) -> bool

Returns true if the depth bounds used by this pipeline are dynamic.

fn has_dynamic_stencil_compare_mask(&self) -> bool

Returns true if the stencil compare masks used by this pipeline are dynamic.

fn has_dynamic_stencil_write_mask(&self) -> bool

Returns true if the stencil write masks used by this pipeline are dynamic.

fn has_dynamic_stencil_reference(&self) -> bool

Returns true if the stencil references used by this pipeline are dynamic.

Loading content...
Provided methods

fn subpass(self) -> Subpass<Self> where
Self: Sized,

Returns the subpass this graphics pipeline is rendering to.

Loading content...