Trait vulkano::pipeline::vertex::VertexSource

pub unsafe trait VertexSource<L> {
    fn decode(
        &self, 
        _: L
    ) -> (Vec<Box<dyn BufferAccess + Send + Sync>>, usize, usize);
}

Extension trait of VertexDefinition. The L parameter is an acceptable vertex source for this vertex definition.

Required methods

fn decode(
&self,
_: L
) -> (Vec<Box<dyn BufferAccess + Send + Sync>>, usize, usize)

Checks and returns the list of buffers with offsets, number of vertices and number of instances.

Loading content...