Enum vulkano::pipeline::viewport::ViewportsState

pub enum ViewportsState {
    Fixed {
        data: Vec<(Viewport, Scissor)>,
    },
    DynamicViewports {
        scissors: Vec<Scissor>,
    },
    DynamicScissors {
        viewports: Vec<Viewport>,
    },
    Dynamic {
        num: u32,
    },
}

List of viewports and scissors that are used when creating a graphics pipeline object.

Note that the number of viewports and scissors must be the same.

Variants
Fixed

The state is known in advance.

Fields

data: Vec<(Viewport, Scissor)>

State of the viewports and scissors.

DynamicViewports

The state of scissors is known in advance, but the state of viewports is dynamic and will bet set when drawing.

Note that the number of viewports and scissors must be the same.

Fields

scissors: Vec<Scissor>

State of the scissors.

DynamicScissors

The state of viewports is known in advance, but the state of scissors is dynamic and will bet set when drawing.

Note that the number of viewports and scissors must be the same.

Fields

viewports: Vec<Viewport>

State of the viewports

Dynamic

The state of both the viewports and scissors is dynamic and will be set when drawing.

Fields

num: u32

Number of viewports and scissors.

Methods

impl ViewportsState

fn dynamic_viewports(&self) -> bool

Returns true if the state of the viewports is dynamic.

fn dynamic_scissors(&self) -> bool

Returns true if the state of the scissors is dynamic.

fn num_viewports(&self) -> u32

Returns the number of viewports and scissors.