Struct vulkano::sampler::Sampler

pub struct Sampler {
    // some fields omitted
}

Describes how to retreive data from an image within a shader.

Methods

impl Sampler

fn new(device: &Arc<Device>, mag_filter: Filter, min_filter: Filter, mipmap_mode: MipmapMode, address_u: SamplerAddressMode, address_v: SamplerAddressMode, address_w: SamplerAddressMode, mip_lod_bias: f32, max_anisotropy: f32, min_lod: f32, max_lod: f32) -> Result<Arc<Sampler>, SamplerCreationError>

Creates a new Sampler with the given behavior.

Panic

Panicks if max_anisotropy < 1.0.

Panicks if min_lod > max_lod.

fn unnormalized(device: &Arc<Device>, filter: Filter, address_u: UnnormalizedSamplerAddressMode, address_v: UnnormalizedSamplerAddressMode) -> Result<Arc<Sampler>, SamplerCreationError>

Creates a sampler with unnormalized coordinates. This means that texture coordinates won't range between 0.0 and 1.0 but use plain pixel offsets.

Using an unnormalized sampler adds a few restrictions:

It can only be used with non-array 1D or 2D images.

It can only be used with images with a single mipmap.

Projection and offsets can't be used by shaders. Only the first mipmap can be accessed.