Struct vulkano::shader::ShaderModule

pub struct ShaderModule {
    // some fields omitted
}

Contains SPIR-V code with one or more entry points.

Note that it is advised to wrap around a ShaderModule with a struct that is different for each shader.

Methods

impl ShaderModule

unsafe fn new(device: &Arc<Device>, spirv: &[u8]) -> Result<Arc<ShaderModule>, OomError>

Builds a new shader module from SPIR-V.

Safety

The SPIR-V code is not validated.

The SPIR-V code may require some features that are not enabled. This isn't checked by this function either.

unsafe fn vertex_shader_entry_point<'a, V, L>(&'a self, name: &'a CStr, attributes: Vec<(u32, Cow<'static, str>)>) -> VertexShaderEntryPoint<'a, V, L>

unsafe fn fragment_shader_entry_point<'a, F, L>(&'a self, name: &'a CStr) -> FragmentShaderEntryPoint<'a, F, L>

Gets access to an entry point contained in this module.

This is purely a logical operation. It returns a struct that represents the entry point but doesn't actually do anything.

Safety

The user must check that the entry point exists in the module, as this is not checked by Vulkan.

Calling this function also determines the template parameters associated to the EntryPoint struct. Therefore care must be taken that the values there are correct.