Struct vulkano::swapchain::display::Display

pub struct Display {
    // some fields omitted
}

Represents a monitor connected to a physical device.

Methods

impl Display

fn enumerate_raw(device: &PhysicalDevice) -> Result<IntoIter<Display>, OomError>

See the docs of enumerate().

fn enumerate(device: &PhysicalDevice) -> IntoIter<Display>

Enumerates all the displays that are available on a given physical device.

Panic

Panics if the device or host ran out of memory.

fn name(&self) -> &str

Returns the name of the display.

fn physical_device(&self) -> PhysicalDevice

Returns the physical device that was used to create this display.

fn physical_resolution(&self) -> [u32; 2]

Returns the physical resolution of the display.

fn display_modes_raw(&self) -> Result<IntoIter<DisplayMode>, OomError>

See the docs of display_modes().

fn display_modes(&self) -> IntoIter<DisplayMode>

Returns a list of all modes available on this display.

Panic

Panics if the device or host ran out of memory.