Struct vulkano::swapchain::display::DisplayPlane

pub struct DisplayPlane {
    // some fields omitted
}

?

Methods

impl DisplayPlane

fn enumerate_raw(device: &PhysicalDevice) -> Result<IntoIter<DisplayPlane>, OomError>

See the docs of enumerate().

fn enumerate(device: &PhysicalDevice) -> IntoIter<DisplayPlane>

Enumerates all the display planes that are available on a given physical device.

Panic

Panics if the device or host ran out of memory.

fn physical_device(&self) -> PhysicalDevice

Returns the physical device that was used to create this display.

fn supports(&self, display: &Display) -> bool

Returns true if this plane supports the given display.