Struct vulkano::swapchain::Display

pub struct Display {
    // some fields omitted
}

Represents a monitor connected to a physical device.

Methods

impl Display

fn enumerate(device: &PhysicalDevice) -> Result<IntoIter<Display>, OomError>

Enumerates all the displays that are available on a given physical device.

fn name(&self) -> &str

Returns the name of the display.

fn physical_resolution(&self) -> [u32; 2]

Returns the physical resolution of the display.

fn display_modes(&self) -> Result<IntoIter<DisplayMode>, OomError>

Returns a list of all modes available on this display.