Struct vulkano::sync::Event

pub struct Event { /* fields omitted */ }

Used to block the GPU execution until an event on the CPU occurs.

Note that Vulkan implementations may have limits on how long a command buffer will wait for an event to be signaled, in order to avoid interfering with progress of other clients of the GPU. If the event isn't signaled within these limits, results are undefined and may include device loss.

Methods

impl Event

pub fn from_pool(device: Arc<Device>) -> Result<Event, OomError>

Takes an event from the vulkano-provided event pool. If the pool is empty, a new event will be allocated. Upon drop, the event is put back into the pool.

For most applications, using the event pool should be preferred, in order to avoid creating new events every frame.

pub fn alloc(device: Arc<Device>) -> Result<Event, OomError>

Builds a new event.

pub fn signaled(&self) -> Result<bool, OomError>

Returns true if the event is signaled.

pub fn set_raw(&mut self) -> Result<(), OomError>

See the docs of set().

pub fn set(&mut self)

Changes the Event to the signaled state.

If a command buffer is waiting on this event, it is then unblocked.

Panic

Panics if the device or host ran out of memory.

pub fn reset_raw(&mut self) -> Result<(), OomError>

See the docs of reset().

pub fn reset(&mut self)

Changes the Event to the unsignaled state.

Panic

Panics if the device or host ran out of memory.