Struct vulkano::sync::Fence

pub struct Fence {
    // some fields omitted
}

A fence is used to know when a command buffer submission has finished its execution.

When a command buffer accesses a ressource, you have to ensure that the CPU doesn't access the same ressource simultaneously (except for concurrent reads). Therefore in order to know when the CPU can access a ressource again, a fence has to be used.

Methods

impl Fence

fn new(device: &Arc<Device>) -> Result<Arc<Fence>, OomError>

Builds a new fence.

fn signaled(device: &Arc<Device>) -> Result<Arc<Fence>, OomError>

Builds a new fence already in the "signaled" state.

fn ready(&self) -> Result<bool, OomError>

Returns true if the fence is signaled.

fn wait(&self, timeout_ns: u64) -> Result<(), OomError>

Waits until the fence is signaled, or at least until the number of nanoseconds of the timeout has elapsed.

Returns Ok if the fence is now signaled. Returns Err if the timeout was reached instead.

fn reset(&self)

Resets the fence.

fn multi_reset<'a, I>(iter: I) where I: IntoIterator<Item=&'a Fence>

Resets multiple fences at once.

Panic

Panicks if not all fences belong to the same device.