Struct vulkano::sync::Semaphore

pub struct Semaphore<D = Arc<Device>> where D: SafeDeref<Target=Device> {
    // some fields omitted
}

Used to provide synchronization between command buffers during their execution.

It is similar to a fence, except that it is purely on the GPU side. The CPU can't query a semaphore's status or wait for it to be signaled.

Methods

impl<D> Semaphore<D> where D: SafeDeref<Target=Device>

fn raw(device: D) -> Result<Semaphore<D>, OomError>

See the docs of new().

fn new(device: D) -> Arc<Semaphore<D>>

Builds a new semaphore.

Panic

Panics if the device or host ran out of memory.