Trait vulkano::sync::Resource

pub unsafe trait Resource {
    fn sharing_mode(&self) -> &SharingMode;

    fn requires_fence(&self) -> bool { ... }
    fn requires_semaphore(&self) -> bool { ... }
}

Base trait for objects that can be used as resources and must be synchronized.

Required Methods

fn sharing_mode(&self) -> &SharingMode

Returns in which queue family or families this resource can be used.

Provided Methods

fn requires_fence(&self) -> bool

Returns true if the gpu_access function should be passed a fence.

fn requires_semaphore(&self) -> bool

Returns true if the gpu_access function should be passed a semaphore.