Struct vulkano::buffer::device_local::DeviceLocalBuffer

pub struct DeviceLocalBuffer<T: ?Sized, A = Arc<StdMemoryPool>> where A: MemoryPool {
    // some fields omitted
}

Buffer whose content is accessible by the CPU.

Methods

impl<T> DeviceLocalBuffer<T>

fn new<'a, I>(device: &Arc<Device>, usage: &Usage, queue_families: I) -> Result<Arc<DeviceLocalBuffer<T>>, OomError> where I: IntoIterator<Item=QueueFamily<'a>>

Builds a new buffer. Only allowed for sized data.

impl<T> DeviceLocalBuffer<[T]>

fn array<'a, I>(device: &Arc<Device>, len: usize, usage: &Usage, queue_families: I) -> Result<Arc<DeviceLocalBuffer<[T]>>, OomError> where I: IntoIterator<Item=QueueFamily<'a>>

Builds a new buffer. Can be used for arrays.

impl<T: ?Sized> DeviceLocalBuffer<T>

unsafe fn raw<'a, I>(device: &Arc<Device>, size: usize, usage: &Usage, queue_families: I) -> Result<Arc<DeviceLocalBuffer<T>>, OomError> where I: IntoIterator<Item=QueueFamily<'a>>

Builds a new buffer without checking the size.

Safety

You must ensure that the size that you pass is correct for T.

impl<T: ?Sized, A> DeviceLocalBuffer<T, A> where A: MemoryPool

fn device(&self) -> &Arc<Device>

Returns the device used to create this buffer.

fn queue_families(&self) -> Vec<QueueFamily>

Returns the queue families this buffer can be used on.