Struct vulkano::buffer::BufferSlice

pub struct BufferSlice<'a, T: ?Sized, B: 'a> {
    // some fields omitted
}

A subpart of a buffer.

This object doesn't correspond to any Vulkan object. It exists for API convenience.

Example

Creating a slice:

use vulkano::buffer::BufferSlice;
                                                        unsafe { std::mem::uninitialized() };
let _slice = BufferSlice::from(&buffer);

Selecting a slice of a buffer that contains [T]:

use vulkano::buffer::BufferSlice;
                                                        unsafe { std::mem::uninitialized() };
let _slice = BufferSlice::from(&buffer).slice(12 .. 14).unwrap();

Methods

impl<'a, T: ?Sized, B: 'a> BufferSlice<'a, T, B>

fn buffer(&self) -> &'a Arc<B>

Returns the buffer that this slice belongs to.

fn offset(&self) -> usize

Returns the offset of that slice within the buffer.

fn size(&self) -> usize

Returns the size of that slice in bytes.

unsafe fn slice_custom<F, R: ?Sized>(self, f: F) -> BufferSlice<'a, R, B> where F: for<'r> FnOnce(&'r T) -> &'r R

Builds a slice that contains an element from inside the buffer.

This method builds an object that represents a slice of the buffer. No actual operation is performed.

Example

TODO

Safety

The object whose reference is passed to the closure is uninitialized. Therefore you must not access the content of the object.

You must return a reference to an element from the parameter. The closure must not panic.

impl<'a, T, B: 'a> BufferSlice<'a, [T], B>

fn len(&self) -> usize

Returns the number of elements in this slice.

fn index(self, index: usize) -> Option<BufferSlice<'a, T, B>>

Reduces the slice to just one element of the array.

Returns None if out of range.

fn slice(self, range: Range<usize>) -> Option<BufferSlice<'a, [T], B>>

Reduces the slice to just a range of the array.

Returns None if out of range.