Module vulkano::command_buffer::pool

In the Vulkan API, command buffers must be allocated from command pools.

A command pool holds and manages the memory of one or more command buffers. If you destroy a command pool, all of its command buffers are automatically destroyed.

In vulkano, creating a command buffer requires passing an implementation of the CommandPool trait. By default vulkano will use the StandardCommandPool struct, but you can implement this trait yourself by wrapping around the UnsafeCommandPool type.

Re-exports

pub use self::standard::StandardCommandPool;

Modules

standard

Structs

UnsafeCommandPool
Low-level implementation of a command pool.

UnsafeCommandPoolAlloc
Opaque type that represents a command buffer allocated from a pool.

UnsafeCommandPoolAllocIter
Iterator for newly-allocated command buffers.

Enums

CommandPoolTrimError
Error that can happen when trimming command pools.

Traits

CommandPool
Types that manage the memory of command buffers.

CommandPoolAlloc
A command buffer allocated from a pool that has finished being recorded.

CommandPoolBuilderAlloc
A command buffer allocated from a pool and that can be recorded.