Struct vulkano::command_buffer::PrimaryCommandBufferBuilderInlineDraw

pub struct PrimaryCommandBufferBuilderInlineDraw<P = Arc<StandardCommandPool>> where P: CommandPool {
    // some fields omitted
}

Object that you obtain when calling draw_inline or next_subpass_inline.

Methods

impl<P> PrimaryCommandBufferBuilderInlineDraw<P> where P: CommandPool

fn draw<V, L, Pv, Pl, Rp, Pc>(self, pipeline: &Arc<GraphicsPipeline<Pv, Pl, Rp>>, vertices: V, dynamic: &DynamicState, sets: L, push_constants: &Pc) -> PrimaryCommandBufferBuilderInlineDraw<P> where Pv: VertexSource<V> + 'static, Pl: PipelineLayout + 'static + Send + Sync, Rp: 'static + Send + Sync, L: DescriptorSetsCollection + Send + Sync, Pc: 'static + Clone + Send + Sync

Calls vkCmdDraw.

fn draw_indexed<'a, V, L, Pv, Pl, Rp, I, Ib, Ibb, Pc>(self, pipeline: &Arc<GraphicsPipeline<Pv, Pl, Rp>>, vertices: V, indices: Ib, dynamic: &DynamicState, sets: L, push_constants: &Pc) -> PrimaryCommandBufferBuilderInlineDraw<P> where Pv: 'static + VertexSource<V> + Send + Sync, Pl: 'static + PipelineLayout + Send + Sync, Rp: 'static + Send + Sync, Ib: Into<BufferSlice<'a, [I], Ibb>>, I: 'static + Index, Ibb: Buffer + 'static + Send + Sync, L: DescriptorSetsCollection + Send + Sync, Pc: 'static + Clone + Send + Sync

Calls vkCmdDrawIndexed.

fn next_subpass_inline(self) -> PrimaryCommandBufferBuilderInlineDraw<P>

Switches to the next subpass of the current renderpass.

This function is similar to draw_inline on the builder.

Panic

Panics if no more subpasses remain.

fn next_subpass_secondary(self) -> PrimaryCommandBufferBuilderSecondaryDraw<P>

Switches to the next subpass of the current renderpass.

This function is similar to draw_secondary on the builder.

Panic

Panics if no more subpasses remain.

fn draw_end(self) -> PrimaryCommandBufferBuilder<P>

Finish drawing this renderpass and get back the builder.

Panic

Panics if not at the last subpass.