Enum vulkano::command_buffer::sys::Kind

pub enum Kind<'a, R: 'a, F: 'a> {
    Primary,
    Secondary,
    SecondaryRenderPass {
        subpass: Subpass<'a, R>,
        framebuffer: Option<&'a F>,
    },
}

Determines the kind of command buffer that we want to create.

Variants
Primary

A primary command buffer can execute all commands and can call secondary command buffers.

Secondary

A secondary command buffer can execute all dispatch and transfer operations, but not drawing operations.

SecondaryRenderPass

A secondary command buffer within a render pass can only call draw operations that can be executed from within a specific subpass.

Fields

subpass: Subpass<'a, R>

Which subpass this secondary command buffer can be called from.

framebuffer: Option<&'a F>

The framebuffer object that will be used when calling the command buffer. This parameter is optional and is an optimization hint for the implementation.