Struct vulkano::command_buffer::sys::UnsafeCommandBufferBuilder

pub struct UnsafeCommandBufferBuilder<P> where P: CommandPool {
    // some fields omitted
}

Methods

impl<P> UnsafeCommandBufferBuilder<P> where P: CommandPool

fn new<R, Rf>(pool: P, kind: Kind<R, Rf>, flags: Flags) -> Result<UnsafeCommandBufferBuilder<P>, OomError> where R: RenderPass + 'static + Send + Sync, Rf: RenderPass + 'static + Send + Sync

Creates a new builder.

unsafe fn already_allocated<R, Rf>(pool: P, cmd: AllocatedCommandBuffer, kind: Kind<R, Rf>, flags: Flags) -> Result<UnsafeCommandBufferBuilder<P>, OomError> where R: RenderPass + 'static + Send + Sync, Rf: RenderPass + 'static + Send + Sync

Creates a new command buffer builder from an already-allocated command buffer.

Safety

The allocated command buffer must belong to the pool and must not be used anywhere else in the code for the duration of this command buffer.

fn build(self) -> Result<UnsafeCommandBuffer<P>, OomError>

Finishes building the command buffer.

fn pool(&self) -> &P

Returns the pool used to create this command buffer builder.

fn device(&self) -> &Arc<Device>

Returns the device this command buffer belongs to.

fn is_secondary(&self) -> bool

Returns true if this is a secondary command buffer.

unsafe fn fill_buffer<T: ?Sized, B>(&mut self, buffer: BufferSlice<T, B>, data: u32) where B: Buffer

Fills a buffer by repeating a 32 bits data.

This is similar to the memset function in C/C++.

Panic

Panicks if the buffer was not created with the same device as this command buffer.

Safety

The buffer must be kept alive and must be properly synchronized while this command buffer runs.

The queue family must support graphics or compute operations.

Type safety is not checked.

The offset must be a multiple of four.

The size must be a multiple of four, or must point to the end of the buffer.

The buffer must have been created with the "transfer_dest" usage.

Must be called outside of a render pass.

unsafe fn update_buffer<T: ?Sized, B, D: ?Sized>(&mut self, buffer: BufferSlice<T, B>, data: &D) where B: Buffer, D: Copy + 'static

Fills a buffer with some data.

The actual size that is copied is the minimum between the size of the slice and the size of the data.

Panic

Panicks if the buffer was not created with the same device as this command buffer.

Safety

The buffer must be kept alive and must be properly synchronized while this command buffer runs.

Type safety is not checked.

The offset and size must be multiples of four.

The size must be less than or equal to 65536 bytes (ie. 64kB).

The buffer must have been created with the "transfer_dest" usage.

Must be called outside of a render pass.

unsafe fn copy_buffer<Bs, Bd, I>(&mut self, src: &Arc<Bs>, dest: &Arc<Bd>, regions: I) where Bs: Buffer, Bd: Buffer, I: IntoIterator<Item=BufferCopyRegion>

Copies data from a source buffer to a destination buffer.

This is similar to the memcpy function in C/C++.

Automatically filters out empty regions.

Panic

Panicks if one of the buffers was not created with the same device as this command buffer.

Safety

The buffers must be kept alive and must be properly synchronized while this command buffer runs.

Type safety is not checked.

The source buffer must have been created with the "transfer_src" usage.

The destination buffer must have been created with the "transfer_dest" usage.

Must be called outside of a render pass.

The offsets and size of the regions must be in range.

fn pipeline_barrier(&mut self, barrier: PipelineBarrierBuilder)

Adds a pipeline barrier to the command buffer.

This function itself is not unsafe, but creating a pipeline barrier builder is.