Module vulkano::descriptor::descriptor

Structs

DescriptorBufferDesc

DescriptorDesc
Describes a single descriptor.

DescriptorImageDesc

ShaderStages
Describes which shader stages have access to a descriptor.

Enums

DescriptorDescTy
Describes the content and layout of each array element of a descriptor.

DescriptorImageDescArray

DescriptorImageDescDimensions

DescriptorType
Describes what kind of resource may later be bound to a descriptor.