Module vulkano::descriptor::descriptor_set

Reexports

pub use self::collection::DescriptorSetsCollection;

Modules

collection

Structs

DescriptorPool
Pool from which descriptor sets are allocated from.

DescriptorWrite
Represents a single write entry to a descriptor set.

UnsafeDescriptorSet
Low-level descriptor set.

UnsafeDescriptorSetLayout
Describes to the Vulkan implementation the layout of all descriptors within a descriptor set.

Traits

DescriptorSet
Trait for objects that contain a collection of resources that will be accessible by shaders.

DescriptorSetDesc
Trait for objects that describe the layout of the descriptors of a set.

TrackedDescriptorSet
Extension trait for descriptor sets so that it can be used with the standard commands list interface.

TrackedDescriptorSetFinished

TrackedDescriptorSetState