Trait vulkano::descriptor::descriptor_set::collection::DescriptorSetsCollection

pub unsafe trait DescriptorSetsCollection {
    type ListIter: ExactSizeIterator<Item=Arc<DescriptorSet>>;
    type SetsIter: ExactSizeIterator<Item=Self::DescIter>;
    type DescIter: ExactSizeIterator<Item=DescriptorDesc>;
    fn list(&self) -> Self::ListIter;
    fn description(&self) -> Self::SetsIter;
}

A collection of descriptor set objects.

Associated Types

type ListIter: ExactSizeIterator<Item=Arc<DescriptorSet>>

An iterator that produces the list of descriptor set objects contained in this collection.

type SetsIter: ExactSizeIterator<Item=Self::DescIter>

An iterator that produces the description of the list of sets.

type DescIter: ExactSizeIterator<Item=DescriptorDesc>

An iterator that produces the description of a set.

Required Methods

fn list(&self) -> Self::ListIter

Returns the list of descriptor set objects of this collection.

fn description(&self) -> Self::SetsIter

Produces a description of the sets, as if it was a layout.