Module vulkano::descriptor::pipeline_layout

Describes the layout of descriptors and push constants used by a pipeline.

This module contains all the structs and traits related to describing the layout of descriptors and push constants used by the shaders of a graphics or compute pipeline.

The layout itself only describes the descriptors and push constants, and does not contain the content of the push constants or the actual list of resources that are going to be available through the descriptors. Push constants are set when you submit a draw command, and the list of resources is set by creating descriptor set objects and passing these sets when you submit a draw command.

Modules

custom_pipeline_macro

Structs

EmptyPipeline
Implementation of PipelineLayout for an empty pipeline.

EmptyPipelineDesc
Description of an empty pipeline layout.

UnsafePipelineLayout
Low-level struct that represents the layout of the resources available to your shaders.

Enums

UnsafePipelineLayoutCreationError
Error that can happen when creating an instance.

Traits

PipelineLayout
Trait for objects that describe the layout of the descriptors and push constants of a pipeline.

PipelineLayoutDesc
Trait for objects that describe the layout of the descriptors and push constants of a pipeline.

PipelineLayoutPushConstantsCompatible
Traits that allow determining whether

PipelineLayoutSetsCompatible
Traits that allow determining whether

PipelineLayoutSuperset
Traits that allow determining whether a pipeline layout is a superset of another one.