Trait vulkano::descriptor::pipeline_layout::PipelineLayoutDesc

pub unsafe trait PipelineLayoutDesc {
    type SetsIter: Iterator<Item=Self::DescIter>;
    type DescIter: Iterator<Item=DescriptorDesc>;
    fn descriptors_desc(&self) -> Self::SetsIter;
}

Trait for objects that describe the layout of the descriptors and push constants of a pipeline.

Associated Types

type SetsIter: Iterator<Item=Self::DescIter>

Iterator that describes descriptor sets.

type DescIter: Iterator<Item=DescriptorDesc>

Iterator that describes individual descriptors.

Required Methods

fn descriptors_desc(&self) -> Self::SetsIter

Describes the layout of the descriptors of the pipeline.