Trait vulkano::format::PossibleFloatOrCompressedFormatDesc

pub unsafe trait PossibleFloatOrCompressedFormatDesc: FormatDesc {
    fn is_float_or_compressed(&self) -> bool;
}

Trait for types that can possibly describe a float or compressed attachment.

Required Methods

fn is_float_or_compressed(&self) -> bool

Returns true if the format is a float or compressed format.