Function vulkano::framebuffer::ensure_image_view_compatible

pub fn ensure_image_view_compatible<Rp: ?Sized, I: ?Sized>(
    render_pass: &Rp, 
    attachment_num: usize, 
    image: &I
) -> Result<(), IncompatibleRenderPassAttachmentError> where
    Rp: RenderPassDesc,
    I: ImageViewAccess, 

Checks whether the given image view is allowed to be the nth attachment of the given render pass.

Panic

Panics if the attachment number is out of range.