Struct vulkano::framebuffer::Framebuffer

pub struct Framebuffer<Rp, A> { /* fields omitted */ }

Contains a render pass and the image views that are attached to it.

Creating a framebuffer is done by calling Framebuffer::start, which returns a FramebufferBuilder object. You can then add the framebuffer attachments one by one by calling add(image). When you are done, call build().

Both the add and the build functions perform various checks to make sure that the number of images is correct and that each image is compatible with the attachment definition in the render pass.

use vulkano::framebuffer::Framebuffer;

// let render_pass: Arc<_> = ...;
let framebuffer = Framebuffer::start(render_pass.clone())
    .add(my_image).unwrap()
    .build().unwrap();

Just like render pass objects implement the RenderPassAbstract trait, all framebuffer objects implement the FramebufferAbstract trait. This means that you can cast any Arc<Framebuffer<..>> into an Arc<FramebufferAbstract + Send + Sync> for easier storage.

Framebuffer dimensions

If you use Framebuffer::start() to create a framebuffer then vulkano will automatically make sure that all the attachments have the same dimensions, as this is the most common situation.

Alternatively you can also use with_intersecting_dimensions, in which case the dimensions of the framebuffer will be the intersection of the dimensions of all attachments, or with_dimensions if you want to specify exact dimensions. If you use with_dimensions, you are allowed to attach images that are larger than these dimensions.

If the dimensions of the framebuffer don't match the dimensions of one of its attachment, then only the top-left hand corner of the image will be drawn to.

Methods

impl<Rp> Framebuffer<Rp, ()>

pub fn start(render_pass: Rp) -> FramebufferBuilder<Rp, ()>

Starts building a framebuffer.

pub fn with_intersecting_dimensions(
render_pass: Rp
) -> FramebufferBuilder<Rp, ()>

Starts building a framebuffer. The dimensions of the framebuffer will automatically be the intersection of the dimensions of all the attachments.

pub fn with_dimensions(
render_pass: Rp,
dimensions: [u32; 3]
) -> FramebufferBuilder<Rp, ()>

Starts building a framebuffer.

impl<Rp, A> Framebuffer<Rp, A>

pub fn dimensions(&self) -> [u32; 3]

Returns the width, height and layers of this framebuffer.

pub fn width(&self) -> u32

Returns the width of the framebuffer in pixels.

pub fn height(&self) -> u32

Returns the height of the framebuffer in pixels.

pub fn layers(&self) -> u32

Returns the number of layers (or depth) of the framebuffer.

pub fn device(&self) -> &Arc<Device>

Returns the device that was used to create this framebuffer.

pub fn render_pass(&self) -> &Rp

Returns the renderpass that was used to create this framebuffer.