Struct vulkano::framebuffer::Framebuffer

pub struct Framebuffer<L> {
    // some fields omitted
}

Contains the list of images attached to a render pass.

This is a structure that you must pass when you start recording draw commands in a command buffer.

A framebuffer can be used alongside with any other render pass object as long as it is compatible with the render pass that his framebuffer was created with. You can determine whether two renderpass objects are compatible by calling is_compatible_with.

Methods

impl<L> Framebuffer<L>

fn new<A>(render_pass: &Arc<L>, dimensions: [u32; 3], attachments: A) -> Result<Arc<Framebuffer<L>>, FramebufferCreationError> where L: RenderPass + RenderPassAttachmentsList<A>

Builds a new framebuffer.

The attachments parameter depends on which RenderPass implementation is used.

fn is_compatible_with<R>(&self, render_pass: &Arc<R>) -> bool where R: RenderPass, L: RenderPass + RenderPassCompatible<R>

Returns true if this framebuffer can be used with the specified renderpass.

fn dimensions(&self) -> [u32; 3]

Returns the width, height and layers of this framebuffer.

fn width(&self) -> u32

Returns the width of the framebuffer in pixels.

fn height(&self) -> u32

Returns the height of the framebuffer in pixels.

fn layers(&self) -> u32

Returns the number of layers (or depth) of the framebuffer.

fn device(&self) -> &Arc<Device>

Returns the device that was used to create this framebuffer.

fn render_pass(&self) -> &Arc<L>

Returns the renderpass that was used to create this framebuffer.

fn attachments(&self) -> &[(Arc<ImageView>, Arc<Image>, ImageLayout, ImageLayout)]

Returns all the resources attached to that framebuffer.