Trait vulkano::framebuffer::RenderPassDesc

pub unsafe trait RenderPassDesc {
    type AttachmentsIter: ExactSizeIterator<Item=LayoutAttachmentDescription>;
    type PassesIter: ExactSizeIterator<Item=LayoutPassDescription>;
    type DependenciesIter: ExactSizeIterator<Item=LayoutPassDependencyDescription>;
    fn attachments(&self) -> Self::AttachmentsIter;
    fn passes(&self) -> Self::PassesIter;
    fn dependencies(&self) -> Self::DependenciesIter;

    fn num_subpasses(&self) -> u32 { ... }
    fn num_color_attachments(&self, subpass: u32) -> Option<u32> { ... }
    fn num_samples(&self, subpass: u32) -> Option<u32> { ... }
    fn has_depth_stencil_attachment(&self, subpass: u32) -> Option<(bool, bool)> { ... }
    fn has_depth(&self, subpass: u32) -> Option<bool> { ... }
    fn has_writable_depth(&self, subpass: u32) -> Option<bool> { ... }
    fn has_stencil(&self, subpass: u32) -> Option<bool> { ... }
    fn has_writable_stencil(&self, subpass: u32) -> Option<bool> { ... }
}

Associated Types

type AttachmentsIter: ExactSizeIterator<Item=LayoutAttachmentDescription>

Iterator returned by the attachments method.

type PassesIter: ExactSizeIterator<Item=LayoutPassDescription>

Iterator returned by the passes method.

type DependenciesIter: ExactSizeIterator<Item=LayoutPassDependencyDescription>

Iterator returned by the dependencies method.

Required Methods

fn attachments(&self) -> Self::AttachmentsIter

Returns an iterator that describes the list of attachments of this render pass.

fn passes(&self) -> Self::PassesIter

Returns an iterator that describes the list of passes of this render pass.

fn dependencies(&self) -> Self::DependenciesIter

Returns an iterator that describes the list of inter-pass dependencies of this render pass.

Provided Methods

fn num_subpasses(&self) -> u32

Returns the number of subpasses within the render pass.

fn num_color_attachments(&self, subpass: u32) -> Option<u32>

Returns the number of color attachments in a subpass. Returns None if out of range.

fn num_samples(&self, subpass: u32) -> Option<u32>

Returns the number of samples of the attachments of a subpass. Returns None if out of range or if the subpass has no attachment. TODO: return an enum instead?

fn has_depth_stencil_attachment(&self, subpass: u32) -> Option<(bool, bool)>

Returns a tuple whose first element is true if there's a depth attachment, and whose second element is true if there's a stencil attachment. Returns None if out of range.

fn has_depth(&self, subpass: u32) -> Option<bool>

Returns true if a subpass has a depth attachment or a depth-stencil attachment.

fn has_writable_depth(&self, subpass: u32) -> Option<bool>

Returns true if a subpass has a depth attachment or a depth-stencil attachment whose layout is not DepthStencilReadOnlyOptimal.

fn has_stencil(&self, subpass: u32) -> Option<bool>

Returns true if a subpass has a stencil attachment or a depth-stencil attachment.

fn has_writable_stencil(&self, subpass: u32) -> Option<bool>

Returns true if a subpass has a stencil attachment or a depth-stencil attachment whose layout is not DepthStencilReadOnlyOptimal.