Module vulkano::image

Images storage (1D, 2D, 3D, arrays, etc.).

An image is a location in memory whose purpose is to store multi-dimensional data. Its most common usage is to store a 2D array of color pixels (in other words an image in the everyday language), but it can also be used to store arbitrary data.

High-level wrappers

The vulkano library provides high-level wrappers around images that are specialized depending on the way you are going to use the image:

An AttachmentImage can be used when you want to draw to an image.

An ImmutableImage stores data which never need be changed after the initial upload, like a texture.

If are a beginner, you are strongly encouraged to use one of these wrappers.

Low-level informations

To be written.

Reexports

pub use self::sys::ImageCreationError;

pub use self::sys::Layout;

pub use self::sys::Usage;

pub use self::traits::Image;

pub use self::traits::ImageView;

Modules

attachment
Image whose purpose is to be used as a framebuffer attachment.

immutable

swapchain

sys
Low-level implementation of images and images views.

traits

Structs

Swizzle

Enums

ComponentSwizzle

MipmapsCount
Specifies how many mipmaps must be allocated.