Struct vulkano::image::immutable::ImmutableImage

pub struct ImmutableImage<F, A = PotentialDedicatedAllocation<StdMemoryPoolAlloc>> { /* fields omitted */ }

Image whose purpose is to be used for read-only purposes. You can write to the image once, but then you must only ever read from it.

Implementations

impl<F> ImmutableImage<F>

pub fn new<'a, I>(
device: Arc<Device>,
dimensions: Dimensions,
format: F,
queue_families: I
) -> Result<Arc<ImmutableImage<F>>, ImageCreationError> where
F: FormatDesc,
I: IntoIterator<Item = QueueFamily<'a>>,

👎 Deprecated:
use ImmutableImage::uninitialized instead

pub fn with_mipmaps<'a, I, M>(
device: Arc<Device>,
dimensions: Dimensions,
format: F,
mipmaps: M,
queue_families: I
) -> Result<Arc<ImmutableImage<F>>, ImageCreationError> where
F: FormatDesc,
I: IntoIterator<Item = QueueFamily<'a>>,
M: Into<MipmapsCount>,

👎 Deprecated:
use ImmutableImage::uninitialized instead

pub fn uninitialized<'a, I, M>(
device: Arc<Device>,
dimensions: Dimensions,
format: F,
mipmaps: M,
usage: ImageUsage,
layout: ImageLayout,
queue_families: I
) -> Result<(Arc<ImmutableImage<F>>, ImmutableImageInitialization<F>), ImageCreationError> where
F: FormatDesc,
I: IntoIterator<Item = QueueFamily<'a>>,
M: Into<MipmapsCount>,

Builds an uninitialized immutable image.

Returns two things: the image, and a special access that should be used for the initial upload to the image.

pub fn from_iter<P, I>(
iter: I,
dimensions: Dimensions,
format: F,
queue: Arc<Queue>
) -> Result<(Arc<Self>, CommandBufferExecFuture<NowFuture, AutoCommandBuffer>), ImageCreationError> where
P: Send + Sync + Clone + 'static,
F: FormatDesc + AcceptsPixels<P> + 'static + Send + Sync,
I: ExactSizeIterator<Item = P>,
Format: AcceptsPixels<P>,

Construct an ImmutableImage from the contents of iter.

TODO: Support mipmaps

pub fn from_buffer<B, P>(
source: B,
dimensions: Dimensions,
format: F,
queue: Arc<Queue>
) -> Result<(Arc<Self>, CommandBufferExecFuture<NowFuture, AutoCommandBuffer>), ImageCreationError> where
B: BufferAccess + TypedBufferAccess<Content = [P]> + 'static + Clone + Send + Sync,
P: Send + Sync + Clone + 'static,
F: FormatDesc + AcceptsPixels<P> + 'static + Send + Sync,
Format: AcceptsPixels<P>,

Construct an ImmutableImage containing a copy of the data in source.

TODO: Support mipmaps

impl<F, A> ImmutableImage<F, A>

pub fn dimensions(&self) -> Dimensions

Returns the dimensions of the image.

pub fn mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the image.