Module vulkano::image::sys

Low-level implementation of images and images views.

This module contains low-level wrappers around the Vulkan image and image view types. All other image or image view types of this library, and all custom image or image view types that you create must wrap around the types in this module.

Structs

LinearLayout
Describes the memory layout of an image with linear tiling.

UnsafeImage
A storage for pixels or arbitrary data.

UnsafeImageView

Usage
Describes how an image is going to be used. This is not an optimization.

Enums

Dimensions

ImageCreationError
Error that can happen when creating an instance.

Layout