Struct vulkano::image::sys::UnsafeImage

pub struct UnsafeImage {
    // some fields omitted
}

A storage for pixels or arbitrary data.

Safety

This type is not just unsafe but very unsafe. Don't use it directly.

You must manually bind memory to the image with bind_memory. The memory must respect the requirements returned by new.

The memory that you bind to the image must be manually kept alive.

The queue family ownership must be manually enforced.

The usage must be manually enforced.

The image layout must be manually enforced and transitionned.

Methods

impl UnsafeImage

unsafe fn new<'a, Mi, I>(device: &Arc<Device>, usage: &Usage, format: Format, dimensions: Dimensions, num_samples: u32, mipmaps: Mi, sharing: Sharing<I>, linear_tiling: bool, preinitialized_layout: bool) -> Result<(UnsafeImage, MemoryRequirements), ImageCreationError> where Mi: Into<MipmapsCount>, I: Iterator<Item=u32>

Creates a new image and allocates memory for it.

Panic

Panics if one of the dimensions is 0.

Panics if the number of mipmaps is 0.

Panics if the number of samples is 0.

unsafe fn from_raw(device: &Arc<Device>, handle: u64, usage: u32, format: Format, dimensions: Dimensions, samples: u32, mipmaps: u32) -> UnsafeImage

Creates an image from a raw handle. The image won't be destroyed.

This function is for example used at the swapchain's initialization.

unsafe fn bind_memory(&self, memory: &DeviceMemory, offset: usize) -> Result<(), OomError>

fn device(&self) -> &Arc<Device>

fn format(&self) -> Format

fn mipmap_levels(&self) -> u32

fn dimensions(&self) -> Dimensions

fn samples(&self) -> u32

unsafe fn color_linear_layout(&self, mip_level: u32) -> LinearLayout

Queries the layout of an image in memory. Only valid for images with linear tiling.

This function is only valid for images with a color format. See the other similar functions for the other aspects.

The layout is invariant for each image. However it is not cached, as this would waste memory in the case of non-linear-tiling images. You are encouraged to store the layout somewhere in order to avoid calling this semi-expensive function at every single memory access.

Note that while Vulkan allows querying the array layers other than 0, it is redundant as you can easily calculate the position of any layer.

Panic

Panics if the mipmap level is out of range.

Safety

The image must not have a depth, stencil or depth-stencil format.

The image must have been created with linear tiling.

unsafe fn depth_linear_layout(&self, mip_level: u32) -> LinearLayout

Same as color_linear_layout, except that it retreives the depth component of the image.

Panic

Panics if the mipmap level is out of range.

Safety

The image must have a depth or depth-stencil format.

The image must have been created with linear tiling.

unsafe fn stencil_linear_layout(&self, mip_level: u32) -> LinearLayout

Same as color_linear_layout, except that it retreives the stencil component of the image.

Panic

Panics if the mipmap level is out of range.

Safety

The image must have a stencil or depth-stencil format.

The image must have been created with linear tiling.

fn supports_blit_source(&self) -> bool

Returns true if the image can be used as a source for blits.

fn supports_blit_destination(&self) -> bool

Returns true if the image can be used as a destination for blits.