Module vulkano::image::traits

Structs

AccessRange

GpuAccessResult

PipelineBarrierRequest
Requests that a pipeline barrier is created.

PipelineMemoryBarrierRequest
Requests that a memory barrier is created as part of the pipeline barrier.

SubmitInfos

Transition

Traits

AttachmentImageView

CommandBufferState
Trait for objects that represent the state of the image in a command buffer.

CommandListState
Trait for objects that represent the state of a slice of the image in a list of commands.

Image
Trait for types that represent images.

ImageClearValue
Extension trait for images. Checks whether the value T can be used as a clear value for the given image.

ImageContent

ImageView
Trait for types that represent image views.

TrackedImage
Extension trait for Image. Types that implement this can be used in a StdCommandBuffer.