Struct vulkano::instance::ApplicationInfo

pub struct ApplicationInfo<'a> {
    pub application_name: Option<Cow<'a, str>>,
    pub application_version: Option<Version>,
    pub engine_name: Option<Cow<'a, str>>,
    pub engine_version: Option<Version>,
}

Information that can be given to the Vulkan driver so that it can identify your application.

Fields
application_name: Option<Cow<'a, str>>

Name of the application.

application_version: Option<Version>

An opaque number that contains the version number of the application.

engine_name: Option<Cow<'a, str>>

Name of the engine used to power the application.

engine_version: Option<Version>

An opaque number that contains the version number of the engine.

Methods

impl<'a> ApplicationInfo<'a>

fn from_cargo_toml() -> ApplicationInfo<'a>

Builds an ApplicationInfo from the information gathered by Cargo.

Panic

Panics if the required environment variables are missing, which happens if the project wasn't built by Cargo.