Module vulkano::memory

GPU-visible memory allocation.

Reexports

pub use self::pool::MemoryPool;

Modules

pool

Structs

CpuAccess
Object that can be used to read or write the content of a MappedDeviceMemory.

DeviceMemory
Represents memory that has been allocated.

MappedDeviceMemory
Represents memory that has been allocated and mapped in CPU accessible space.

MemoryRequirements
Represents requirements expressed by the Vulkan implementation.

Traits

Content
Trait for types of data that can be mapped.