Struct vulkano::memory::DeviceMemory

pub struct DeviceMemory<D = Arc<Device>> where D: SafeDeref<Target=Device> {
    // some fields omitted
}

Represents memory that has been allocated.

Methods

impl<D> DeviceMemory<D> where D: SafeDeref<Target=Device>

fn alloc(device: &D, memory_type: &MemoryType, size: usize) -> Result<DeviceMemory<D>, OomError> where D: Clone

Allocates a chunk of memory from the device.

Some platforms may have a limit on the maximum size of a single allocation. For example, certain systems may fail to create allocations with a size greater than or equal to 4GB.

Panic

Panics if size is 0.

Panics if memory_type doesn't belong to the same physical device as device.

fn alloc_and_map(device: &D, memory_type: &MemoryType, size: usize) -> Result<MappedDeviceMemory<D>, OomError> where D: Clone

Allocates a chunk of memory and maps it.

Panic

Panics if memory_type doesn't belong to the same physical device as device.

Panics if the memory type is not host-visible.

fn memory_type(&self) -> MemoryType

Returns the memory type this chunk was allocated on.

fn size(&self) -> usize

Returns the size in bytes of that memory chunk.

fn device(&self) -> &Device

Returns the device associated with this allocation.