Module vulkano::pipeline

Describes a graphical or compute operation.

In Vulkan before you can add a draw or a compute command to a command buffer, you have to create a pipeline object that describes it.

There are two kinds of pipelines:

ComputePipelines, for compute operations that reads/write raw pixels from images and buffers.

GraphicsPipelines, for graphical operations.

All the sub-modules of this module (with the exception of cache) correspond to the various steps of pipelines.

Modules

blend
Defines how the color output of the fragment shader is written to the attachment.

cache
Cache the pipeline objects to disk for faster reloads.

depth_stencil
Depth and stencil operations description.

input_assembly
Assembling vertices into primitives.

multisample
State of multisampling.

raster
Stage when triangles are turned into pixels.

shader
Stage of a graphics pipeline.

vertex

viewport
Viewports and scissor boxes.

Structs

ComputePipeline
A pipeline object that describes to the Vulkan implementation how it should perform compute operations.

GraphicsPipeline
Defines how the implementation should perform a draw operation.

GraphicsPipelineParams
Description of a GraphicsPipeline.

GraphicsPipelineParamsTess
Additional parameters if you use tessellation.

Enums

GraphicsPipelineCreationError
Error that can happen when creating a graphics pipeline.