Struct vulkano::pipeline::cache::PipelineCache

pub struct PipelineCache {
    // some fields omitted
}

Opaque cache that contains pipeline objects.

Methods

impl PipelineCache

unsafe fn with_data(device: &Arc<Device>, initial_data: &[u8]) -> Result<Arc<PipelineCache>, OomError>

Builds a new pipeline cache from existing data.

The data must have been previously obtained with get_data.

fn empty(device: &Arc<Device>) -> Result<Arc<PipelineCache>, OomError>

Builds a new empty pipeline cache.

fn merge<'a, I>(&self, pipelines: I) -> Result<(), OomError> where I: IntoIterator<Item=&'a &'a Arc<PipelineCache>>

Merges other pipeline caches into this one.

Panic

Panics if self is included in the list of other pipelines.

fn get_data(&self) -> Result<Vec<u8>, OomError>

Obtains the data from the cache.

This data can be stored and then reloaded and passed to PipelineCache::new.