Struct vulkano::pipeline::ComputePipeline

pub struct ComputePipeline<Pl> {
    // some fields omitted
}

A pipeline object that describes to the Vulkan implementation how it should perform compute operations.

The template parameter contains the descriptor set to use with this pipeline.

Methods

impl<Pl> ComputePipeline<Pl>

fn new<Css, Csl>(device: &Arc<Device>, pipeline_layout: &Arc<Pl>, shader: &ComputeShaderEntryPoint<Css, Csl>, specialization: &Css) -> Result<Arc<ComputePipeline<Pl>>, ComputePipelineCreationError> where Pl: PipelineLayout + PipelineLayoutSuperset<Csl>, Csl: PipelineLayoutDesc, Css: SpecializationConstants

Builds a new ComputePipeline.

fn device(&self) -> &Arc<Device>

Returns the Device this compute pipeline was created with.

fn layout(&self) -> &Arc<Pl>

Returns the pipeline layout used in this compute pipeline.