Struct vulkano::pipeline::GraphicsPipeline

pub struct GraphicsPipeline<VertexDefinition, Layout, RenderP> { /* fields omitted */ }

Defines how the implementation should perform a draw operation.

This object contains the shaders and the various fixed states that describe how the implementation should perform the various operations needed by a draw command.

Methods

impl GraphicsPipeline<(), (), ()>

pub fn start<'a>(
) -> GraphicsPipelineBuilder<BufferlessDefinition, EmptyEntryPointDummy, (), EmptyEntryPointDummy, (), EmptyEntryPointDummy, (), EmptyEntryPointDummy, (), EmptyEntryPointDummy, (), ()>

Starts the building process of a graphics pipeline. Returns a builder object that you can fill with the various parameters.

impl<Mv, L, Rp> GraphicsPipeline<Mv, L, Rp>

pub fn vertex_definition(&self) -> &Mv

Returns the vertex definition used in the constructor.

pub fn device(&self) -> &Arc<Device>

Returns the device used to create this pipeline.

impl<Mv, L, Rp> GraphicsPipeline<Mv, L, Rp> where
L: PipelineLayoutAbstract,

pub fn layout(&self) -> &L

Returns the pipeline layout used in the constructor.

impl<Mv, L, Rp> GraphicsPipeline<Mv, L, Rp> where
Rp: RenderPassDesc,

pub fn subpass(&self) -> Subpass<&Rp>

Returns the pass used in the constructor.

impl<Mv, L, Rp> GraphicsPipeline<Mv, L, Rp>

pub fn render_pass(&self) -> &Rp

Returns the render pass used in the constructor.

pub fn has_dynamic_line_width(&self) -> bool

Returns true if the line width used by this pipeline is dynamic.

pub fn num_viewports(&self) -> u32

Returns the number of viewports and scissors of this pipeline.

pub fn has_dynamic_viewports(&self) -> bool

Returns true if the viewports used by this pipeline are dynamic.

pub fn has_dynamic_scissors(&self) -> bool

Returns true if the scissors used by this pipeline are dynamic.

pub fn has_dynamic_depth_bounds(&self) -> bool

Returns true if the depth bounds used by this pipeline are dynamic.

pub fn has_dynamic_stencil_compare_mask(&self) -> bool

Returns true if the stencil compare masks used by this pipeline are dynamic.

pub fn has_dynamic_stencil_write_mask(&self) -> bool

Returns true if the stencil write masks used by this pipeline are dynamic.

pub fn has_dynamic_stencil_reference(&self) -> bool

Returns true if the stencil references used by this pipeline are dynamic.