Struct vulkano::pipeline::GraphicsPipeline

pub struct GraphicsPipeline<VertexDefinition, Layout, RenderP> {
    // some fields omitted
}

The template parameter contains the descriptor set to use with this pipeline, and the renderpass layout.

Methods

impl<Vdef, L, Rp> GraphicsPipeline<Vdef, L, Rp> where L: PipelineLayout, Rp: RenderPass + RenderPassDesc

fn new<'a, Vsp, Vi, Vo, Vl, Fs, Fi, Fo, Fl>(device: &Arc<Device>, params: GraphicsPipelineParams<'a, Vdef, Vsp, Vi, Vo, Vl, (), (), (), (), Fs, Fi, Fo, Fl, L, Rp>) -> Result<Arc<GraphicsPipeline<Vdef, L, Rp>>, GraphicsPipelineCreationError> where Vdef: VertexDefinition<Vi>, L: PipelineLayout + PipelineLayoutSuperset<Vl> + PipelineLayoutSuperset<Fl>, Vl: PipelineLayoutDesc, Fl: PipelineLayoutDesc, Fi: ShaderInterfaceDefMatch<Vo>, Fo: ShaderInterfaceDef, Vo: ShaderInterfaceDef, Rp: RenderPassSubpassInterface<Fo>

Builds a new graphics pipeline object.

fn with_geometry_shader<'a, Vsp, Vi, Vo, Vl, Gsp, Gi, Go, Gl, Fs, Fi, Fo, Fl>(device: &Arc<Device>, params: GraphicsPipelineParams<'a, Vdef, Vsp, Vi, Vo, Vl, Gsp, Gi, Go, Gl, Fs, Fi, Fo, Fl, L, Rp>) -> Result<Arc<GraphicsPipeline<Vdef, L, Rp>>, GraphicsPipelineCreationError> where Vdef: VertexDefinition<Vi>, L: PipelineLayout + PipelineLayoutSuperset<Vl> + PipelineLayoutSuperset<Fl>, Vl: PipelineLayoutDesc, Fl: PipelineLayoutDesc, Gi: ShaderInterfaceDefMatch<Vo>, Vo: ShaderInterfaceDef, Fi: ShaderInterfaceDefMatch<Go> + ShaderInterfaceDefMatch<Vo>, Fo: ShaderInterfaceDef, Go: ShaderInterfaceDef, Rp: RenderPassSubpassInterface<Fo>

Builds a new graphics pipeline object with a geometry shader.

impl<Mv, L, Rp> GraphicsPipeline<Mv, L, Rp>

fn vertex_definition(&self) -> &Mv

Returns the vertex definition used in the constructor.

impl<Mv, L, Rp> GraphicsPipeline<Mv, L, Rp> where L: PipelineLayout

fn layout(&self) -> &Arc<L>

Returns the pipeline layout used in the constructor.

impl<Mv, L, Rp> GraphicsPipeline<Mv, L, Rp> where Rp: RenderPass + RenderPassDesc

fn render_pass(&self) -> &Arc<Rp>

Returns the render pass used in the constructor.

fn subpass(&self) -> Subpass<Rp>

Returns the pass used in the constructor.

impl<Mv, L, Rp> GraphicsPipeline<Mv, L, Rp>

fn has_dynamic_line_width(&self) -> bool

Returns true if the line width used by this pipeline is dynamic.

fn num_viewports(&self) -> u32

Returns the number of viewports and scissors of this pipeline.

fn has_dynamic_viewports(&self) -> bool

Returns true if the viewports used by this pipeline are dynamic.

fn has_dynamic_scissors(&self) -> bool

Returns true if the scissors used by this pipeline are dynamic.

fn has_dynamic_depth_bounds(&self) -> bool

Returns true if the depth bounds used by this pipeline are dynamic.

fn has_dynamic_stencil_compare_mask(&self) -> bool

Returns true if the stencil compare masks used by this pipeline are dynamic.

fn has_dynamic_stencil_write_mask(&self) -> bool

Returns true if the stencil write masks used by this pipeline are dynamic.

fn has_dynamic_stencil_reference(&self) -> bool

Returns true if the stencil references used by this pipeline are dynamic.