Trait vulkano::pipeline::vertex::VertexMember

pub unsafe trait VertexMember {
    fn format() -> (VertexMemberTy, usize);
}

Trait for data types that can be used as vertex members. Used by the impl_vertex! macro.

Required methods

fn format() -> (VertexMemberTy, usize)

Returns the format and array size of the member.

Loading content...