Struct vulkano::sync::Event

pub struct Event {
    // some fields omitted
}

Used to block the GPU execution until an event on the CPU occurs.

Note that Vulkan implementations may have limits on how long a command buffer will wait for an event to be signaled, in order to avoid interfering with progress of other clients of the GPU. If the event isn't signaled within these limits, results are undefined and may include device loss.

Methods

impl Event

fn raw(device: &Arc<Device>) -> Result<Event, OomError>

See the docs of new().

fn new(device: &Arc<Device>) -> Arc<Event>

Builds a new event.

Panic

Panics if the device or host ran out of memory.

fn signaled(&self) -> Result<bool, OomError>

Returns true if the event is signaled.

fn set_raw(&mut self) -> Result<(), OomError>

See the docs of set().

fn set(&mut self)

Changes the Event to the signaled state.

If a command buffer is waiting on this event, it is then unblocked.

Panic

Panics if the device or host ran out of memory.

fn reset_raw(&mut self) -> Result<(), OomError>

See the docs of reset().

fn reset(&mut self)

Changes the Event to the unsignaled state.

Panic

Panics if the device or host ran out of memory.